### 2020-12-26 #2929 @whedon
[PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models
**Submitting author:** @alemermartinez (<a href="http://orcid.org/0000-0002-1893-2039">Alejandra Martínez</a>)
**Repository:** <a href="https://github.com/msalibian/RBF" target ="_blank">https://github.com/msalibian/RBF</a>
**Version:** v2.0.1
