### 2020-12-26 #2929 @danielskatz
👋 @alemermartinez - I'm unsure if this fits JOSS, so the editors will now discuss if it meets JOSS's scope for review, specifically to check the [size/scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) of the software. This normally would take about a week, but might take longer due to the holidays.
