### 2020-12-29 #2929 @Kevin-Mattheus-Moerman
@alemermartinez I just reviewed your paper as part of the scope check. I noted the following typo: `additve`. If you make any changes to the paper that you want to update you can call `@whedon generate pdf` here to update the paper.
