### 2020-12-29 #2929 @alemermartinez
________________________________
De: Kevin Mattheus Moerman <notifications@github.com>
Enviado: martes, 29 de diciembre de 2020 06:49
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

@alemermartinez<https://github.com/alemermartinez> I just reviewed your paper as part of the scope check. I noted the following typo: additve. If you make any changes to the paper that you want to update you can call @whedon generate pdf here to update the paper.
