### 2020-12-29 #2929 @alemermartinez
________________________________
De: whedon <notifications@github.com>
Enviado: sábado, 26 de diciembre de 2020 11:35
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

Submitting author: @alemermartinez<https://github.com/alemermartinez> (Alejandra Martínez<http://orcid.org/0000-0002-1893-2039>)
Repository: <https://github.com/msalibian/RBF> https://github.com/msalibian/RBF
Version: v2.0.1
Editor: Pending
Reviewer: Pending
Managing EiC: Daniel S. Katz

⚠️ JOSS reduced service mode ⚠️

Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in our blog post<https://blog.joss.theoj.org/2020/05/reopening-joss>.

Author instructions

Thanks for submitting your paper to JOSS @alemermartinez<https://github.com/alemermartinez>. Currently, there isn't an JOSS editor assigned to your paper.

@alemermartinez<https://github.com/alemermartinez> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people<https://bit.ly/joss-reviewers> have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Editor instructions

The JOSS submission bot @whedon<https://github.com/whedon> is here to help you find and assign reviewers and start the main review. To find out what @whedon<https://github.com/whedon> can do for you type:
