### 2020-12-29 #2929 @alemermartinez
________________________________
De: whedon <notifications@github.com>
Enviado: martes, 29 de diciembre de 2020 11:20
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

# List of editor GitHub usernames

EDITORIAL TASKS
