### 2021-01-04 #2929 @mikldk
@alemermartinez - thank you for the submission.

Do you have any suggestions for reviewers ([here](https://bit.ly/joss-reviewers) is a list of potential reviewers)?

The `README.Rmd` (and thereby `README.md`) seems to have a conflict that must be resolved (indicated by `<<<<<<< HEAD` and other similar markers).
