### 2021-01-06 #2929 @alemermartinez
Dear Mikkel.

I have deleted both README files (and their corresponding Figures) from the "paper" branch, along with the DESCRIPTION file.

I also note that some files in the software branch of the repository (README and DESCRIPTION) have also been flagged as a conflict. However, I'd rather not make changes to the software repository, since those files play a specific role there. What can I do to fix the warning messages coming from the software branch?

With respect to potential reviewers for the paper, the following seem appropriate:

@fabian-s
@mcavs
@kellieotto
@mmrabe
@prdm0
@pedro-teles-fonseca

Kind regards,

Alejandra

-------- Mensaje original --------
De: Mikkel Meyer Andersen <notifications@github.com>
Fecha: 4/1/21 06:06 (GMT-03:00)
A: openjournals/joss-reviews <joss-reviews@noreply.github.com>
CC: Alejandra Martinez <ale_m_martinez@hotmail.com>, Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

@alemermartinez<https://github.com/alemermartinez> - thank you for the submission.

Do you have any suggestions for reviewers (here<https://bit.ly/joss-reviewers> is a list of potential reviewers)?

The README.Rmd (and thereby README.md) seems to have a conflict that must be resolved (indicated by <<<<<<< HEAD and other similar markers).
