### 2021-01-06 #2929 @mikldk
@alemermartinez - please remember not to mention potential reviewers, i.e. leave out the `@` prefix. I took the liberty and changed your message above.

Also, I am not sure about the structure of your repository. Is it on purpose that you have a separate branch for JOSS (`joss`) instead of just the `master` branch?

I definitely think that the conflicts in `README` should be fixed no matter what.
