### 2021-01-09 #2929 @alemermartinez
I am sorry for having mentioned the possible reviewers.

In order to separate the paper from the software, I was suggested to have the paper on a specific branch of the software repository and not in the master one.
This submission is a re-submission because in the first attempt the paper was stored in a different GitHub repository from the one containing the software.

https://github.com/openjournals/joss-reviews/issues/2910#issuecomment-747612707

Here, Argon says "Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default)."
This is what we tried to do.

On the other hand and just to understand, what is the conflict with the README files? There must be just one file with extension .md or .Rmd?
Anyway, if you consider that is necessary, I can also remove all README files from the master branch too.

Kind regards,

Alejandra

________________________________
De: Mikkel Meyer Andersen <notifications@github.com>
Enviado: miércoles, 6 de enero de 2021 11:04
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

@alemermartinez<https://github.com/alemermartinez> - please remember not to mention potential reviewers, i.e. leave out the @ prefix. I took the liberty and changed your message above.

Also, I am not sure about the structure of your repository. Is it on purpose that you have a separate branch for JOSS (joss) instead of just the master branch?

I definitely think that the conflicts in README should be fixed no matter what.
