### 2021-01-11 #2929 @mikldk
@alemermartinez ,

I suggest to have the software and the paper in the same repo in the same branch. Potentially, you can have the paper (paper.md + misc) in its own branch (which is just as fine if you prefer that). So that you totally have one copy of the software and one copy of the paper, potentially across multiple branches. Now you have one copy of the software (the package) in each branch. That may be fine if you are used to git with pull requests etc., but it can also be more confusing if you're not.

So I suggest that you from the JOSS branch remove the software and only have the paper.

It is fine that you have README.Rmd that is used to generate README.md. The conflict is marked by text like this:

```
<<<<<<< HEAD Robust backfitting
```

You can read more about it e.g. [here](https://www.atlassian.com/git/tutorials/using-branches/merge-conflicts) and [here](https://docs.github.com/en/free-pro-team@latest/github/collaborating-with-issues-and-pull-requests/resolving-a-merge-conflict-using-the-command-line).
