### 2021-01-11 bc8c764
Delete NAMESPACE
Deleted from joss branch
