+++ NAMESPACE
@@ -1,23 +1,0 @@
-# Generated by roxygen2: do not edit by hand
-
-S3method(deviance,backf)
-S3method(deviance,backf.cl)
-S3method(deviance,backf.rob)
-S3method(fitted,backf)
-S3method(fitted.values, backf)
-S3method(formula,backf)
-S3method(plot,backf)
-S3method(predict,backf)
-S3method(print,backf)
-S3method(residuals,backf)
-S3method(summary,backf)
-S3method(summary,backf.cl)
-S3method(summary,backf.rob)
-export(backf.cl)
-export(backf.rob)
-export(k.epan)
-export(psi.huber)
-export(psi.tukey)
-import(graphics)
-import(stats)
-useDynLib(RBF, .registration = TRUE)
