### 2021-01-11 1efd5f2
Delete man directory
Deleted from joss branch
