+++ man/formula.backf.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RBF-fn.R
-\name{formula.backf}
-\alias{formula.backf}
-\title{Additive model formula}
-\usage{
-\method{formula}{backf}(x, ...)
-}
-\arguments{
-\item{x}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}
-
-\item{...}{additional other arguments. Currently ignored.}
-}
-\value{
-A model formula.
-}
-\description{
-Description of the additive model formula extracted from an object of class \code{backf}.
-}
-\author{
-Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
-}
