+++ man/predict.backf.Rd
@@ -1,25 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RBF-fn.R
-\name{predict.backf}
-\alias{predict.backf}
-\title{Fitted values for objects of class \code{backf}.}
-\usage{
-\method{predict}{backf}(object, ...)
-}
-\arguments{
-\item{object}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}
-
-\item{...}{additional other arguments. Currently ignored.}
-}
-\value{
-A vector of fitted values.
-}
-\description{
-This function returns the fitted values given the covariates of
-the original sample under an additive model using the classical or
-robust backfitting approach computed with \code{\link{backf.cl}} or
-\code{\link{backf.rob}}.
-}
-\author{
-Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
-}
