+++ man/k.epan.Rd
@@ -1,29 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RBF-fn.R
-\name{k.epan}
-\alias{k.epan}
-\title{Epanechnikov kernel}
-\usage{
-k.epan(x)
-}
-\arguments{
-\item{x}{a vector of real numbers}
-}
-\value{
-A vector of the same length as \code{x} where each entry is
-\code{0.75 * (1 - x^2)} if \code{x < 1} and 0 otherwise.
-}
-\description{
-This function evaluates an Epanechnikov kernel
-}
-\details{
-This function evaluates an Epanechnikov kernel
-}
-\examples{
-x <- seq(-2, 2, length=10)
-k.epan(x)
-
-}
-\author{
-Matias Salibian-Barrera, \email{matias@stat.ubc.ca}, Alejandra Martinez
-}
