+++ man/residuals.backf.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RBF-fn.R
-\name{residuals.backf}
-\alias{residuals.backf}
-\title{Residuals for objects of class \code{backf}}
-\usage{
-\method{residuals}{backf}(object, ...)
-}
-\arguments{
-\item{object}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}
-
-\item{...}{additional other arguments. Currently ignored.}
-}
-\value{
-A vector of residuals.
-}
-\description{
-This function returns the residuals of the fitted additive model using
-the classical or robust backfitting estimators, as computed with \code{\link{backf.cl}} or
-\code{\link{backf.rob}}.
-}
-\author{
-Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
-}
