### 2021-01-11 c840d63
Delete src directory
Deleted from joss branch
