+++ src/RBF-fn.c
@@ -1,703 +1,0 @@
-// #include <stdio.h>
-// #include <stdlib.h>
-// #include <math.h>
-
-#include <R.h>
-#include "RBF.h"
-
-#define ZERO 1e-10
-#define TOL_INVERSE ZERO
-#define min(a,b) ((a)<(b)?(a):(b))
-
-double psi_huber_w( double r, double k) {
-double sal, aux1, aux2;
-aux1 = fabs(r);
-aux2 = k/aux1;
-// sal = min(aux2, 1.0);
-if( aux2 < 1 ) { sal = aux2; }
-else { sal = 1;}
-return( sal );
-}
-
-
-double psi_tukey_w( double r, double k) {
-double sal=0.0, aux1;
-aux1 = fabs(r/k);
-if( aux1 < 1.0 ) {
-	sal = (1+aux1) * (1+aux1) * (1-aux1) * (1-aux1);
-};
-return( sal );
-}
-
-// MAL
-// double psi_tukey_w( double r, double k) {
-// double sal, aux1, aux2;
-// aux1 = fabs(r/k);
-// aux2 = (1+aux1) * (1+aux1) * (1-aux1) * (1-aux1);
-// if( aux2 < 1 ) { sal = aux2; }
-// else { sal = 0;}
-// return( sal );
-// }
-
-
-
-
-//31
-//Nucleo de Epanechnikov
-
-double kepan( double b ) {
-double a, aux;
-a = 0.75 * (1 - (b*b) );
-aux = fabs(b);
-if( aux < 1 ) { return(a); }
-else { return(0); }
-}
-//41
-//Suma de coordenadas de un vector
-
-
-double suma_vec(double *x, int n){
-register int i;
-double sal=0;
-for(i=0;i<n;i++){
-	sal = sal + x[i];
-}
-return(sal);
-}
-
-double norma_dos(double *x, int n){
-register int i;
-double sal=0;
-for(i=0;i<n;i++){
-	sal = sal + (x[i])*(x[i]);
-}
-return(sqrt(sal));
-}
-
-double l2dist(double *x, double *y, int n){
-register int i;
-double sal=0;
-for(i=0;i<n;i++){
-	sal = sal + (x[i] - y[i]) * (x[i] - y[i]);
-}
-return(sqrt(sal));
-}
-
-
-
-
-void kernel_huber_pos(double *punto, double *x, int *nrow, double *y, double *muhat_initial,
- double *ventanas, double *eps, double *sigmahat, double *prob,
- double *kh, int *maxit, double *salida) {
-
-double kepan(double);
-double psi_huber_w(double, double);
-double suma_vec(double*, int);
-register int i;
-int n = *nrow, it=0;
-double corte, muhat, muold, khh;
-double *pesos, *aux1, *aux2, *res;
-
-pesos = (double *) malloc( n * sizeof(double));
-aux1 = (double *) malloc( n * sizeof(double));
-aux2 = (double *) malloc( n * sizeof(double));
-res = (double *) malloc( n * sizeof(double));
-corte = 10 * (*eps);
-khh = *kh;
-muhat = *muhat_initial;
-for(i=0;i<n;i++) {
-     pesos[i] = kepan( (x[i] - *punto) / *ventanas )  / prob[i];
-}
-
-while( (corte > (*eps)) && (it< (*maxit)) ){
-	for(i=0;i<n;i++){
-		res[i] = ( y[i] - muhat) / *sigmahat;
-		aux1[i] = pesos[i] * psi_huber_w(res[i],khh) * y[i];
-		aux2[i] = pesos[i] * psi_huber_w(res[i],khh);
-	};
-	muold = muhat;
[…]
