+++ README.Rmd
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 ---
 title: "Robust backfitting"
 author: "Matias Salibian"
@@ -348,348 +347,3 @@ boxplot(tmspe.ro, tmspe.gam, names=c('Robust', 'Classical'),
 boxplot(mspe.ro, mspe.gam, names=c('Robust', 'Classical'), 
         col=rep(c('tomato3', 'gray80'), 2), main='Non-Trimmed MSPE On "clean" data')
 ```
-=======
----
-title: "Robust backfitting"
-author: "Matias Salibian"
-date: "`r format(Sys.Date())`"
-output: github_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message=FALSE,
-                      fig.width=6, fig.height=6)
-```
-
-## A robust backfitting algorithm
-
-The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF)) 
-implements the robust back-fitting algorithm as proposed by
-Boente, Martinez and Salibian-Barrera  in 
-
-> Boente G, Martinez A, Salibian-Barrera M. (2017) Robust estimators for 
-> additive models using backfitting. Journal of Nonparametric Statistics. Taylor 
-> & Francis; 29, 744-767.
-> [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077)
-
-This repository contains a development version of `RBF`
-which may differ slightly from the one available on CRAN
-(until the CRAN version is updated appropriately). 
-
-The package in this repository can be installed from within `R` by using the following code (assuming the [devtools](https://cran.r-project.org/package=devtools)) package is available:
-```R
-devtools::install_github("msalibian/RBF")
-```
-
-### An example
-
-Here is a (longish) example on how `RBF` works. 
-We use the Air Quality data. The interest is in
-predicting `Ozone` in terms of three explanatory 
-variables: `Solar.R`, `Wind` and `Temp`:
-```{r intro}
-library(RBF)
-data(airquality)
-pairs(airquality[, c('Ozone', 'Solar.R', 'Wind', 'Temp')], 
-      pch=19, col='gray30', cex=1.5)
-```
-
-The following bandwidths were obtained via a robust 
-leave-one-out cross-validation procedure (described in the paper).
-As a robust prediction error measure we use `mu^2 + sigma^2` where
-`mu` and `sigma` are M-estimators of location and scale of
-the prediction errors, respectively. The code is copied below:
-```{r robust.leaveoneout, eval=FALSE}
-# Bandwidth selection with leave-one-out cross-validation
-## Without outliers
-# This takes a long time to compute (approx 380 minutes running
-# R 3.6.1 on an Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz)
-ccs <- complete.cases(airquality)
-aircomplete <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
-a <- c(1/2, 1, 1.5, 2, 2.5, 3)
-h1 <- a * sd(aircomplete[,2])
-h2 <- a * sd(aircomplete[,3])
-h3 <- a * sd(aircomplete[,4])
-hh <- expand.grid(h1, h2, h3)
-nh <- nrow(hh)
-rmspe <- rep(NA, nh)
-jbest <- 0
-cvbest <- +Inf
[…]
