+++ README.md
@@ -1,7 +1,7 @@
 Robust backfitting
 
 Matias Salibian
-2020-07-08
+2021-01-18
 
 ## A robust backfitting algorithm
 
@@ -233,376 +233,6 @@ smoothing parameters for the non-robust fit were re-computed using
 leave-one-out cross-validation on the “clean” data set). Note that now
 both fits (robust and non-robust) are  
 almost identical.
-
-``` r
-# Run the classical backfitting algorithm without outliers
-airclean <- aircomplete[-in.ro, ]
-fit.gam2 <- gam(Ozone ~ lo(Solar.R, span=.7) + lo(Wind, span=.8)+
-                  lo(Temp, span=.3), data=airclean)
-fits2 <- predict(fit.gam2, type='terms')
-# alpha.gam2 <- attr(fits2, 'constant')
-dd2 <- aircomplete[-in.ro, c('Solar.R', 'Wind', 'Temp')]
-for(j in 1:3) {
-  re <- fit.full$yp - fit.full$alpha - rowSums(fit.full$g.matrix[,-j])
-  plot(re ~ x[,j], type='p', pch=20, col='gray45', xlab=colnames(x)[j], ylab='')
-  points(re[in.ro] ~ x[in.ro,j], pch=20, col='red', cex=1.5)
-  oo <- order(dd2[,j])
-  lines(dd2[oo,j], fits2[oo,j], lwd=5, col='magenta')
-  oo <- order(x[,j])
-  lines(x[oo,j], fit.full$g.matrix[oo,j], lwd=5, col='blue')
-}
-```
-
-<img src="README_files/figure-gfm/bothonclean-1.png" width="33%" /><img src="README_files/figure-gfm/bothonclean-2.png" width="33%" /><img src="README_files/figure-gfm/bothonclean-3.png" width="33%" />
-
-Finally, we compare the prediction accuracy obtained with each of the
-fits. Because we are not interested in predicting well any possible
-outliers in the data, we evaluate the quality of the predictions using a
-5%-trimmed mean squared prediction error (effectively measuring the
-prediction accuracy on 95% of the data). We use this alpha-trimmed mean
-squared function:
-
-``` r
-tms <- function(a, alpha=.1) {
-  # alpha is the proportion to trim
-  a2 <- sort(a^2, na.last=NA)
-  n0 <- floor( length(a) * (1 - alpha) )
-  return( mean(a2[1:n0], na.rm=TRUE) )
-}
-```
-
-We use 100 runs of 5-fold CV to compare the 5%-trimmed mean squared
-prediction error of the robust fit and the classical one. Note that the
-bandwidths are kept fixed at their optimal value estimated above.
-
-``` r
-dd <- airquality
-dd <- dd[complete.cases(dd), c('Ozone', 'Solar.R', 'Wind', 'Temp')]
-# 100 runs of K-fold CV
-M <- 100
-# 5-fold
-K <- 5
-n <- nrow(dd)
-# store (trimmed) TMSPE for robust and gam, and also
-tmspe.ro <- tmspe.gam <- vector('numeric', M)
-set.seed(123)
-ii <- (1:n)%%K + 1
-for(runs in 1:M) {
-  tmpro <- tmpgam <- vector('numeric', n)
-  ii <- sample(ii)
-  for(j in 1:K) {
-    fit.full <- backf.rob(Ozone ~ Solar.R + Wind + Temp, 
-                           point=dd[ii==j, -1], windows=bandw, 
-                           epsilon=1e-6, degree=1, type='Tukey', 
-                           subset = (ii!=j), data = dd)
-    tmpro[ ii == j ] <- rowSums(fit.full$prediction) + fit.full$alpha
[…]
