+++ README.md
@@ -1,5 +1,3 @@
-## Robust backfitting
-
 Matias Salibian
 2021-01-18
 
