### 2021-01-23 #2929 @alemermartinez
Dear Mikkel,

It seems we cannot replicate the conflict, probably because we had deleted the repeated files in the joss branch by that moment. For this reason, we have updated all flagged files and deleted all conflict markers.
Command "git status" doesn't show any warning or conflict message, not in the master or the joss branch.

If there is still any conflict, I think that the next step will probably be to remove the joss branch and put the paper in the master branch.

We apologise for the inconvenience.

Kind regards,

Alejandra

________________________________
De: Mikkel Meyer Andersen <notifications@github.com>
Enviado: martes, 12 de enero de 2021 17:20
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

@alemermartinez<https://github.com/alemermartinez>, please also fix the conflict at the master branch as I mentioned. Please do not write here until everything I mentioned is fixed.
