+++ refs.bib
@@ -372,6 +372,7 @@ doi = {10.1214/aos/1176350694}
   edition = {1st},
   publisher = {Wiley},
   address = {New York},
+  doi = {10.1017/cbo9780511618994},
 }
 
 
