### 2021-01-26 #2929 @alemermartinez
________________________________
De: Mikkel Meyer Andersen <notifications@github.com>
Enviado: lunes, 25 de enero de 2021 16:58
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

@alemermartinez<https://github.com/alemermartinez> Thanks. Please check which references you actually use from refs.bib in the paper, and be sure that only those are included in the file and that they all have a DOI.
