### 2021-01-26 #2929 @alemermartinez
Dear Mikkel,

I have just deleted all unnecessary references from the refs.bib file and added the missing DOI.

Kind regards,

Alejandra

________________________________
De: whedon <notifications@github.com>
Enviado: martes, 26 de enero de 2021 12:13
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2929)

OK DOIs

- 10.1080/10485252.2017.1369077 is OK
- 10.1016/0047-259X(89)90023-7 is OK
- 10.2307/2287576 is OK
- 10.1201/9780203753781-6 is OK

MISSING DOIs

- None

INVALID DOIs

- None
