### 2021-01-28 #2992 @whedon
[REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models
**Submitting author:** @alemermartinez (<a href="http://orcid.org/0000-0002-1893-2039">Alejandra Martínez</a>)
**Repository:** <a href="https://github.com/msalibian/RBF" target ="_blank">https://github.com/msalibian/RBF</a>
**Version:** v2.1.0
