### 2021-02-09 #2992 @mcavs
Hi @mmrabe, I have completed review.

I think the manuscript is well-written and easy to read. It would be good to explain the other options of the type argument in the backf.rob function. Also, a short text about their superiority against each other could be added. The fact that the manuscript was illustrated through an example makes me worry about whether the findings are generalizable. I think it is ready for the publication, if this situation can be clarified.
