### 2021-02-10 msalibian/RBF#1 @mmrabe
JOSS review: Fix citations
Refers to JOSS submission openjournals/joss-reviews#2992

Please use proper citation commands, i.e. `[@citation]` instead of `(@citation)` for citations in parentheses. See ll. 24, 34, 42, 47, 51, 62.
