### 2021-02-10 msalibian/RBF#2 @mmrabe
JOSS review: Fix bibliography
Refers to JOSS submission openjournals/joss-reviews#2992

Please check your bibliography.
- TJ Hastie is incoherently named. Please add middle name/initial for Hastie (2019) so that the citations in the paper are correctly displayed. Currently it appears as if T. Hastie and T. J. Hastie were different authors.
- Please add curly braces around capitalized abbreviations in the publication title (esp. for GAM and R) so that they are correctly displayed in the bibliography.
