### 2021-02-10 msalibian/RBF#3 @mmrabe
JOSS review: Typo "cross alidation"
Refers to JOSS submission openjournals/joss-reviews#2992

See l. 61
