### 2021-02-10 msalibian/RBF#4 @mmrabe
JOSS Review: Please revise summary/print functionality
Refers to JOSS submission openjournals/joss-reviews#2992

Please consider revising the `print` methods for your package classes, particularly objects returned by the `backf.*(...)` functions. Currently, when printing the `fit` object as returned by `fit<-backf.*(...)` only a formula is displayed but some more info may be desirable, such as what type of object this is and what it contains.

Moreover, many fitting packages in R return a summary object when calling `summary(fit)`. Please also consider changing your summary functions to return a summary object (e.g., of class `summary.backf`), not just print the output. The printing could be done in a `print` method for that summary object class, such such as a `print.summary.backf(...)` method.

In addition, when printing the summary, output functions such as `cat(...)` that write to `stdout`, not to `stderr`, may be more suitable as such as summary is typically considered expected output, not a warning or message to notify the user of unexpected input or errors.
