### 2021-02-10 msalibian/RBF#5 @mmrabe
JOSS Review: Add vignettes
Refers to JOSS submission openjournals/joss-reviews#2992

Please add the example from your JOSS paper and at least one more as vignettes with a little more detail. If you do so, maybe also consider reducing the detail of the current example and instead summarize findings of the two (or more) examples in the vignettes.
