### 2021-02-10 #2992 @mmrabe
The `RBF` package is an R package for robust backfitting (Boente et al., 2017). Overall, the package works well and the authors provide a good justification and explanation in the paper. I think the paper needs some revisions (esp. with regard to citations/bibliography). Also, for usability the `print` and `summary` methods in the package may need some work but I'm not sure if that should be the scope of this review or may be attended to later on as the package develops. I have added and linked issues to the package repository.

I found your example in the paper very helpful. I'm not a statistician, so I can't confirm whether the algorithm performs exactly what it's supposed to and @mcavs also pointed out that a single example may not be enough to judge that. It may be useful to add at least one more example as a vignette in the package. If you decide to do so, you should of course also consider adding the current example as a vignette. Once the examples are added as vignettes, it may also be worthwhile to only summarize the key findings (e.g., compare the model fits using figures or some statistic) and to refer to the vignettes for more details.

I think this submission is certainly a suitable candidate for publication in JOSS once the open issues have been addressed.
