### 2021-02-10 421ee71
'As.numeric' in windows added
