### 2021-02-10 f308e17
'as.numeric' for 'windows[j]' added
