+++ paper.md
@@ -1,0 +1,358 @@
+---
+title: 'RBF: An R package to compute a robust backfitting estimator
+  for additive models'
+tags:
+- R
+- Backfitting
+- Additive models
+- Robustness
+date: "17 November 2020"
+output:
+  pdf_document: default
+  word_document: default
+  html_document:
+    df_print: paged
+authors:
+- name: Alejandra M. Mart&iacute;nez
+  orcid: 0000-0002-1893-2039
+  affiliation: 1
+- name: Matias Salibian-Barrera
+  orcid: 0000-0003-1873-4611
+  affiliation: 2
+bibliography: refs.bib
+affiliations:
+- name: Departamento de Ciencias Básicas, Universidad Nacional de Luj&aacute;n, Argentina
+  index: 1
+- name: Department of Statistics, University of British Columbia, Canada
+  index: 2
+---
+  
+# Summary
+
+Although highly flexible, non-parametric regression models typically require
+large sample sizes to be estimated reliably, particularly when they include
+many explanatory variables. Additive models provide an alternative that is
+more flexible than linear models, not affected by the curse of
+dimensionality, and also allow the exploration of individual covariate
+effects. Standard algorithms to fit these models can be highly susceptible to
+the presence of a few atypical or outlying observations in the data.
+The ``RBF`` [@RBF] package for R implements the robust estimator for additive
+models of @BoenteMartinezSalibian2017, which can resist the damaging effect
+of outliers in the training set.
+
+# Statement of Need
+
+The purpose of ``RBF`` is to provide a user-friendly implementation of a
+robust kernel-based estimation procedure for additive models that is resistant to
+the presence of potential outliers. 
+
+# Implementation Goals
+
+``RBF`` implements a user interface similar to that of the R package ``gam``
+[@gam], which computes the standard non-robust kernel-based fit for additive
+models using the backfitting algorithm. The ``RBF`` package also includes
+several modeling tools, including functions to produce diagnostic plots,
+obtain fitted values and compute predictions.
+
+# Background
+
+<!-- Nonparametric regression models provide an alternative to parametric regression models such as linear regression models that require fewer and weaker assumptions  for the corresponding analysis to be correct.  In what follows we will assume that 
+$(\textbf{X}_i^\top,Y_i)^\top$, $1\leq i\leq n$, are independent and identically
+distributed random vectors with the same distribution as $(\textbf{X}^\top,Y)^\top$,
+$\textbf{X}\in\mathbb{R}^d$, $Y\in\mathbb{R}$, satisfying the following nonparametric
+regression model:
+$$
+Y \ = \ g(\textbf{X}) \, + \, \epsilon \, ,
+$$
+where the error $\epsilon$ is independent from $\textbf{X}$ and
+$g:\mathbb{R}^d\to\mathbb{R}$ is the unknown regression function to be estimated. Without additional assumptions, fitting this model requires multivariate smoothing
+techniques, which suffer from what is known in the literature as the *curse of dimensionality* [@Stone1985]. Intuitively, the problem is that
+as the number of explanatory variables $d$ increases, neighbourhoods of a point
+$\textbf{x} \in \mathbb{R}^d$ became more sparse.  This results in slower rates of
+convergence for the regression function estimator.  For example, kernel-based estimators
+have a convergence rate of $\sqrt{n \, h_n^d}$, where $h_n \to 0$ is the bandwidth or
[…]
