+++ refs.bib
@@ -1,570 +1,0 @@
-@article{Bianco+Boente:2007,
-author = {Ana Bianco and Graciela Boente},
-title = {Robust estimators under semi-parametric partly linear autoregression: Asymptotic behaviour and bandwidth selection},
-journal = {Journal of Time Series Analysis},
-volume = {28},
-number = {2},
-pages = {274-306},
-year = {2007},
-doi = {10.1111/j.1467-9892.2006.00511.x}
-}
-
-
-
-@book{esl2009,
-    author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
-    title = {The Elements of Statistical Learning -- Data Mining, Inference, and Prediction},
-    year = {2009},
-     publisher={Springer Series in Statistics}
-}
-
-@book{curseofdim,
-author = {Bellman, R.E.},
-year = {1961},
-title = {Adaptive Control Processes}, 
-publisher = {Princeton University Press}
-}
-
-
-@Article{Boente+Martinez:2016,
-  author = {Graciela Boente and Alejandra Mart\'{\i}nez},
-  title = {Estimating additive models with missing responses},
-  journal = {Communications in Statistics: Theory and Methods},
-  year = {2016},
-  volume = {45},
-  number = {2},
-  pages = {413--426},
-  doi = {10.1080/03610926.2013.815780},
-}
-
-@Article{Boente+Martinez:2017,
-  author = {Graciela Boente and Alejandra Mart\'{\i}nez},
-  title = {Marginal integration $M-$estimators for additive models},
-  journal = {TEST},
-  year = {2017},
-  volume = {26},
-  number = {2},
-  pages = {231--260},
-  doi = {10.1007/s11749-016-0508-0},
-}
-
-@Article{BoenteMartinezSalibian2017,
-  author = {Graciela Boente and Alejandra Mart\'{\i}nez and Matias Salibian-Barrera},
-  title = {Robust estimators for additive models using backfitting},
-  journal = {Journal of Nonparametric Statistics},
-  year = {2017},
-  volume = {29},
-  number = {4},
-  pages = {744--767},
-  doi = {10.1080/10485252.2017.1369077},
-}
-
-@Article{Breiman+Friedman:1985,
-  author = {L. Breiman and J. Friedman},
-  title = {Estimating optimal transformations for multiple regression and correlation},
-  journal = {Journal of the American Statistical Association},
-  year = {1985},
-  volume = {80},
-  number = {391},
-  pages = {580--598},
-  doi = {10.1080/01621459.1985.10478157},
-}
-
-@Article{BoenteFraiman1989,
-  author = {Graciela Boente and Ricardo Fraiman},
-  title = {Robust Nonparametric Regression Estimation},
-  journal = {Journal of Multivariate Analysis}, 
-  year = {1989},
-  volume = {29}, 
-  number = {2},
-  pages = {180--198},
-  doi = {10.1016/0047-259X(89)90023-7},
-}
-
-
-@Article{BujaHastieTibshirani1989,
-  author = {A. Buja and T. Hastie and R. Tibshirani},
-  title = {Linear smoothers and additive models (with discussion)},
-  journal = {Annals of Statistics},
-  year = {1989},
-  volume = {17},
-  number = {2},
-  pages = {453--555},
-  doi = {10.1214/aos/1176347115},
-}
-
-@Book{Cameron+Trivedi:2013,
-  author = {A. Colin Cameron and Pravin K. Trivedi},
-  title = {Regression Analysis of Count Data},
-  year = {2013},
-  edition = {2nd},
-  publisher = {Cambridge University Press},
[…]
