+++ DESCRIPTION
@@ -12,6 +12,8 @@ Description: A robust backfitting algorithm for additive models based on (robust
              Boente, Martinez and Salibian-Barrera (2017) <doi:10.1080/10485252.2017.1369077> for
              details. 
 License: GPL (>=3.0)
-RoxygenNote: 6.1.1
+RoxygenNote: 7.1.1
 Encoding: UTF-8
 Imports: stats, graphics
+Suggests: knitr, rmarkdown
+VignetteBuilder: knitr
