### 2021-02-12 #2992 @alemermartinez
Thanks the reviewers for their constructive comments. If that is okay, we will start to incorporate the suggestions.

Kind regards,

Alejandra

________________________________
De: whedon <notifications@github.com>
Enviado: jueves, 11 de febrero de 2021 06:24
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2992)

👋 @mmrabe<https://github.com/mmrabe>, please update us on how your review is going.
