+++ vignettes/Examples.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Examples"
-author: "Martínez y Salibian"
+author: "Martínez and Salibian"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Examples}
@@ -11,23 +11,527 @@ vignette: >
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE)
 ```
+## About this vignette
 
-## R Markdown
+In this vignette, we will show with a toy example and a real data example how a robust backfitting procedure can be easily implemented.
 
-This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.
+The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF)) 
+implements the robust back-fitting algorithm as proposed by
+Boente, Martinez and Salibian-Barrera  in 
 
-When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:
+> Boente G, Martinez A, Salibian-Barrera M. (2017) Robust estimators for 
+> additive models using backfitting. Journal of Nonparametric Statistics. Taylor 
+> & Francis; 29, 744-767.
+> [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077)
 
-```{r cars}
-summary(cars)
+This repository contains a development version of `RBF`
+which may differ slightly from the one available on CRAN
+(until the CRAN version is updated appropriately). 
+
+The package in this repository can be installed from within `R` by using the following code (assuming the [devtools](https://cran.r-project.org/package=devtools)) package is available:
+```R
+devtools::install_github("msalibian/RBF")
 ```
 
-## Including Plots
+Now that the `R` package is downloaded, we can now start to see how this procedure works.
 
-You can also embed plots, for example:
+## Let's play with a toy example
 
-```{r pressure, echo=FALSE}
-plot(pressure)
+The following example corresponds to a 2-dimensional simulated samples with 5% of contaminated responses.
+
+Let begin by defining the additive functions and then generating the simulated sample.
+
+```{r generating data}
+function.g1 <- function(x1) 24*(x1-1/2)^2-2
+function.g2 <- function(x2) 2*pi*sin(pi*x2)-4
+set.seed(140)
+n <- 500
+x1 <- runif(n)
+x2 <- runif(n)
+X <- cbind(x1, x2)
+eps <- rnorm(n,0,sd=0.15)
+regression <- function.g1(x1)+function.g2(x2)
+y <- regression + eps
 ```
 
-Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
[…]
