+++ refs.bib
@@ -12,7 +12,7 @@ doi = {10.1111/j.1467-9892.2006.00511.x}
 
 
 @book{esl2009,
-    author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
+    author = {Trevor J. Hastie and Robert Tibshirani and Jerome Friedman},
     title = {The Elements of Statistical Learning -- Data Mining, Inference, and Prediction},
     year = {2009},
      publisher={Springer Series in Statistics}
@@ -83,7 +83,7 @@ publisher = {Princeton University Press}
 
 
 @Article{BujaHastieTibshirani1989,
-  author = {A. Buja and T. Hastie and R. Tibshirani},
+  author = {A. Buja and T. J. Hastie and R. Tibshirani},
   title = {Linear smoothers and additive models (with discussion)},
   journal = {Annals of Statistics},
   year = {1989},
@@ -211,7 +211,7 @@ doi = {10.1214/aos/1176350694}
 
 
 @Book{HastieTibshirani1990,
-  editor = {T.J. Hastie and R.J. Tibshirani},
+  editor = {T. J. Hastie and R. J. Tibshirani},
   title = {Generalized Additive Models},
   publisher = {Chapman \& Hall},
   year = {1990},
@@ -511,7 +511,7 @@ doi = {10.1214/aos/1176350694}
 
 @Manual{gam,
     title = {gam: Generalized Additive Models},
-    author = {Trevor Hastie},
+    author = {Trevor J. Hastie},
     year = {2019},
     note = {R package version 1.16.1},
     url = {https://CRAN.R-project.org/package=gam}
