+++ refs.bib
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 @article{Bianco+Boente:2007,
 author = {Ana Bianco and Graciela Boente},
 title = {Robust estimators under semi-parametric partly linear autoregression: Asymptotic behaviour and bandwidth selection},
@@ -48,6 +49,8 @@ publisher = {Princeton University Press}
   doi = {10.1007/s11749-016-0508-0},
 }
 
+=======
+>>>>>>> e70290c5f838d87cafceed208cf7883a9b3c3b6c
 @Article{BoenteMartinezSalibian2017,
   author = {Graciela Boente and Alejandra Mart\'{\i}nez and Matias Salibian-Barrera},
   title = {Robust estimators for additive models using backfitting},
@@ -57,17 +60,6 @@ publisher = {Princeton University Press}
   number = {4},
   pages = {744--767},
   doi = {10.1080/10485252.2017.1369077},
-}
-
-@Article{Breiman+Friedman:1985,
-  author = {L. Breiman and J. Friedman},
-  title = {Estimating optimal transformations for multiple regression and correlation},
-  journal = {Journal of the American Statistical Association},
-  year = {1985},
-  volume = {80},
-  number = {391},
-  pages = {580--598},
-  doi = {10.1080/01621459.1985.10478157},
 }
 
 @Article{BoenteFraiman1989,
@@ -82,6 +74,7 @@ publisher = {Princeton University Press}
 }
 
 
+<<<<<<< HEAD
 @Article{BujaHastieTibshirani1989,
   author = {A. Buja and T. J. Hastie and R. Tibshirani},
   title = {Linear smoothers and additive models (with discussion)},
@@ -102,6 +95,8 @@ publisher = {Princeton University Press}
   address = {Cambridge},
 }
 
+=======
+>>>>>>> e70290c5f838d87cafceed208cf7883a9b3c3b6c
 @Book{ChambersClevelandKleinerTukey1983,
   author = {J. M. Chambers and W. S. Cleveland and B. Kleiner and P. A. Tukey},
   title = {Graphical Methods for Data Analysis},
@@ -111,55 +106,6 @@ publisher = {Princeton University Press}
   address = {London},
 }
 
-@Book{Chambers+Hastie:1991,
-  editor = {John M. Chambers and Trevor J. Hastie},
-  title = {Statistical Models in \proglang{S}},
-  publisher = {Chapman \& Hall},
-  year = {1991},
-  address = {London},
-}
-
-@Book{Chen+Hardle+Linton+Severance:1996,
-  author = {R. Chen and W. H\"ardle and O. B. Linton and E. Severance-Lossin},
-  title = {Nonparametric estimation of additive separable regression models},
-  publisher = {Physica-Verlag HD},
-  year = {1996},
-  pages = {453--555},
-  address = {Switzerland},
-  doi = {10.1007/978-3-642-48425-4_18},
-}
-
-
-@Book{deBoor:2001,
-  editor = {Carl de Boor},
-  title = {A Practical Guide to Splines with 32 figures},
-  publisher = {Springer-Verlag},
-  year = {2001},
-  address = {New York},
-}
-
-@Article{Eilers+Marx:1996,
-  author = {Eilers, P.H.C. and Marx, B.D.},
-  title = {Flexible Smoothing with B-splines and Penalties},
-  year = {1996},
-  journal = {Statistical Science},
-  volume = {11},
-  number = {2},
-  pages = {89--121},
-  doi = {10.1214/ss/1038425655},
-}
-
-
-@Article{Fessler:1991,
-  author = {Fessler, J. A.},
-  title = {Nonparametric fixed-interval smoothing with vector splines},
-  year = {1991},
[…]
