+++ refs.bib
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 @article{Bianco+Boente:2007,
 author = {Ana Bianco and Graciela Boente},
 title = {Robust estimators under semi-parametric partly linear autoregression: Asymptotic behaviour and bandwidth selection},
@@ -9,7 +8,6 @@ pages = {274-306},
 year = {2007},
 doi = {10.1111/j.1467-9892.2006.00511.x}
 }
-
 
 
 @book{esl2009,
@@ -49,8 +47,6 @@ publisher = {Princeton University Press}
   doi = {10.1007/s11749-016-0508-0},
 }
 
-=======
->>>>>>> e70290c5f838d87cafceed208cf7883a9b3c3b6c
 @Article{BoenteMartinezSalibian2017,
   author = {Graciela Boente and Alejandra Mart\'{\i}nez and Matias Salibian-Barrera},
   title = {Robust estimators for additive models using backfitting},
@@ -74,7 +70,6 @@ publisher = {Princeton University Press}
 }
 
 
-<<<<<<< HEAD
 @Article{BujaHastieTibshirani1989,
   author = {A. Buja and T. J. Hastie and R. Tibshirani},
   title = {Linear smoothers and additive models (with discussion)},
@@ -95,8 +90,6 @@ publisher = {Princeton University Press}
   address = {Cambridge},
 }
 
-=======
->>>>>>> e70290c5f838d87cafceed208cf7883a9b3c3b6c
 @Book{ChambersClevelandKleinerTukey1983,
   author = {J. M. Chambers and W. S. Cleveland and B. Kleiner and P. A. Tukey},
   title = {Graphical Methods for Data Analysis},
@@ -153,6 +146,3 @@ publisher = {Princeton University Press}
     note = {R package version 2.0.1},
     url = {https://CRAN.R-project.org/package=RBF}
   }
-
-
-
