+++ refs.bib
@@ -62,7 +62,7 @@
 
 @Manual{gam,
     title = {gam: Generalized Additive Models},
-    author = {Trevor J. Hastie},
+    author = {T. J. Hastie},
     year = {2019},
     note = {R package version 1.16.1},
     url = {https://CRAN.R-project.org/package=gam}
