### 2021-02-17 65a5231
New S3 methods
