### 2021-02-17 c74109e
print instead of cat with formula
