### 2021-02-18 7e9bafe
Update paper.md
Additional paragraph
