+++ vignettes/Examples.Rmd
@@ -320,34 +320,8 @@ plot(fit.full, which=1:3)
 ```
 
 <!---
-We now compute and display the 
-classical backfitting fits, with 
-bandwidths chosen via leave-one-out CV:
-```{r gam.loo, eval=FALSE, warning=FALSE}
-library(gam)
-x <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
-a <- c(.3, .4, .5, .6, .7, .8, .9)
-hh <- expand.grid(a, a, a)
-nh <- nrow(hh)
-jbest <- 0
-cvbest <- +Inf
-n <- nrow(x)
-for(i in 1:nh) {
-  fi <- rep(0, n)
-  for(j in 1:n) {
-    tmp <- gam(Ozone ~ lo(Solar.R, span=hh[i,1]) + lo(Wind, span=hh[i,2])
-               + lo(Temp, span=hh[i,3]), data=x, subset=c(-j))
-    fi[j] <- as.numeric(predict(tmp, newdata=x[j,], type='response'))
-  }
-  ss <- mean((x$Ozone - fi)^2)
-  if(ss < cvbest) {
-    jbest <- i
-    cvbest <- ss
-  }
-  print(c(i, ss))
-}
-(hh[jbest,])
-```
+Aca iria cv para gam
+
 --->
 
 The optimal bandwidths are `0.7`, `0.7` and `0.5` for `Solar.R`, 
@@ -433,118 +407,6 @@ for(j in 1:3) {
 }
 ```
 
-<!---
+<!--- Y aca desde que dice Finally
 
-Finally, we compare the prediction accuracy obtained with each
-of the fits. Because we are not interested in predicting well any
-possible outliers in the data, we evaluate the quality of
-the predictions using a 5%-trimmed mean 
-squared prediction error (effectively measuring the prediction 
-accuracy on 95% of the data). We use this alpha-trimmed mean 
-squared function:
-```{r pred1}
-tms <- function(a, alpha=.1) {
-  # alpha is the proportion to trim
-  a2 <- sort(a^2, na.last=NA)
-  n0 <- floor( length(a) * (1 - alpha) )
-  return( mean(a2[1:n0], na.rm=TRUE) )
-}
-```
-
-We use 100 runs of 5-fold CV to compare the 
-5%-trimmed mean squared prediction error of
-the robust fit and the classical one. 
-Note that the bandwidths are kept fixed 
-at their optimal value estimated above. 
-```{r pred2, cache=TRUE, warning=FALSE}
-dd <- airquality
-dd <- dd[complete.cases(dd), c('Ozone', 'Solar.R', 'Wind', 'Temp')]
-# 100 runs of K-fold CV
-M <- 100
-# 5-fold
-K <- 5
-n <- nrow(dd)
-# store (trimmed) TMSPE for robust and gam, and also
-tmspe.ro <- tmspe.gam <- vector('numeric', M)
-set.seed(123)
-ii <- (1:n)%%K + 1
-for(runs in 1:M) {
-  tmpro <- tmpgam <- vector('numeric', n)
-  ii <- sample(ii)
-  for(j in 1:K) {
-    fit.full <- backf.rob(Ozone ~ Solar.R + Wind + Temp, 
-                           point=dd[ii==j, -1], windows=bandw, 
-                           epsilon=1e-6, degree=1, type='Tukey', 
-                           subset = (ii!=j), data = dd)
-    tmpro[ ii == j ] <- rowSums(fit.full$prediction) + fit.full$alpha
-    fit.gam <- gam(Ozone ~ lo(Solar.R, span=.7) + lo(Wind, span=.7)+
[…]
