+++ paper.md
@@ -133,12 +133,12 @@ t^2$, this approach reduces to the standard backfitting.
 
 # Illustration
 
-<span style="color:red">With a real data set we will illustrate how can be used
+\textcolor{red}{With a real data set we will illustrate how can be used
 and it performs the robust backfitting estimate and how it behaves differently from
 the classical approach in the presence of atypical observations. Additional simulated examples with $d=2$ and $d=4$ covariates under different contamination settings can be found in
 @BoenteMartinezSalibian2017. That work also contains a comparison between the robust
-estimates when using the Huber's (``type='Huber'``) and the Tukey's (``type='Tukey'``) loss function and when considering local constant (``degree=0``) or local linear (``degree=1``) smoothers. </span> <span style="color:blue"> Aunque no se si me convence que aparezcan los argumentos antes de que mencionemos
-exactamente que son. Tal vez esta ultima oracion pueda ir mas abajo.</span>
+estimates when using the Huber's (``type='Huber'``) and the Tukey's (``type='Tukey'``) loss function and when considering local constant (``degree=0``) or local linear (``degree=1``) smoothers.} \textcolor{blue}{Aunque no se si me convence que aparezcan los argumentos antes de que mencionemos
+exactamente que son. Tal vez esta ultima oracion pueda ir mas abajo.}
 
 
 The ``airquality`` data set contains 153 daily air quality measurements in
