### 2021-02-22 9e2276d
add formula output to print methods (via calls)
