+++ paper.md
@@ -382,7 +382,7 @@ contains detailed scripts reproducing the data analysis above.
 
 Contributions to this project can be submitted via pull requests on the GitHub repository. 
 Similarly, GitHub issues are the prefered venue to report suggestions and
-problems with the current version of the software and seek support. 
+problems with the current version of the software, and seek support. 
 
 # Acknowledgements
 
