+++ refs.bib
@@ -53,7 +53,7 @@
 
 
 @book{maronna2018robust,
-  title={Robust Statistics: Theory and Methods (with R)},
+  title={Robust Statistics: Theory and Methods (with {R})},
   author={Maronna, Ricardo A. and Martin, R. Douglas and Yohai, Victor J. and Salibian-Barrera, Matias},
   year={2018},
   edition = {2nd},
@@ -61,7 +61,7 @@
 }
 
 @Manual{gam,
-    title = {gam: Generalized Additive Models},
+    title = {{gam}: Generalized Additive Models},
     author = {T. J. Hastie},
     year = {2019},
     note = {R package version 1.16.1},
@@ -70,7 +70,7 @@
 
 
 @Manual{RBF,
-    title = {RBF: Robust Backfitting},
+    title = {{RBF}: Robust Backfitting},
     author = {Matias Salibian-Barrera and Alejandra Mart\'{\i}nez},
     year = {2020},
     note = {R package version 2.0.1},
