+++ refs.bib
@@ -61,7 +61,7 @@
 }
 
 @Manual{gam,
-    title = {{gam}: Generalized Additive Models},
+    title = {gam: Generalized Additive Models},
     author = {T. J. Hastie},
     year = {2019},
     note = {R package version 1.16.1},
@@ -70,7 +70,7 @@
 
 
 @Manual{RBF,
-    title = {{RBF}: Robust Backfitting},
+    title = {RBF: Robust Backfitting},
     author = {Matias Salibian-Barrera and Alejandra Mart\'{\i}nez},
     year = {2020},
     note = {R package version 2.0.1},
