### 2021-02-24 #2992 @alemermartinez
Dear Mikkel and reviewers,

Once again, thank you very much for your constructive comments and suggestions, and also for Mikkel's help with our submission process.

Our revised manuscript and software incorporates all your recommended changes. Specifically: we fixed the references and typos, included community guidelines (lines 115-121), added information on other arguments of the main function (lines 78-91), and incorporated a comment (lines 49-56) about other evidence of the favourable behaviour of our proposal (e.g. a simulated study reported in Boente et al. 2017).
With respect to the software, we have changed the 'summary' and 'print' methods as suggested. We have also started writing a package vignette as suggested by one of the referees, where we plan to include another illustrative example.

Kind regards,

Alejandra

________________________________
De: Mikkel Meyer Andersen <notifications@github.com>
Enviado: martes, 16 de febrero de 2021 10:15
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2992)

@alemermartinez<https://github.com/alemermartinez> Please do. Please check in again within a week's time if possible.
