+++ vignettes/Examples.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
 ```
 ## About this vignette
 
-In this vignette, we will show with a toy example and a real data example how a robust backfitting procedure can be easily implemented.
+In this vignette, we will show with a real data example and some contaminated responses how a robust backfitting procedure can be easily implemented.
 
 The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF)) 
 implements the robust back-fitting algorithm as proposed by
@@ -36,381 +36,77 @@ The package in this repository can be installed from within `R` by using the fol
 ```R
 devtools::install_github("msalibian/RBF")
 ```
-
+and charged in the `R` session by the following command
+```{r library}
+library("RBF")
+```
 Now that the `R` package is downloaded, we can now start to see how this procedure works.
 
-## Let's play with a toy example
+We consider the Boston house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). File `housing.csv` available in this GitHub repository contains $n=506$ observations for the census districts of the Boston metropolitan area. As in this book, we selected the following ten explanatory variables
 
-The following example corresponds to a 2-dimensional simulated samples with 5% of contaminated responses.
+$X_1$: per capita crime rate by town,
 
-Let begin by defining the additive functions and then generating the simulated sample.
+$X_2$: proportion of non-retail business acres per town,
 
-```{r generating data}
-function.g1 <- function(x1) 24*(x1-1/2)^2-2
-function.g2 <- function(x2) 2*pi*sin(pi*x2)-4
-set.seed(140)
-n <- 500
-x1 <- runif(n)
-x2 <- runif(n)
-X <- cbind(x1, x2)
-eps <- rnorm(n,0,sd=0.15)
-regression <- function.g1(x1)+function.g2(x2)
-y <- regression + eps
+$X_3$: nitric oxides concentration (parts per 10 million),
+
+$X_4$: avarage number of rooms per dwelling,
+
+$X_5$: proportion of owner-occupied units built prior to 1940,
+
+$X_6$: weighted distances to five Boston employment centers,
+
+$X_7$: full-value property tax rate per $10,000,
+
+$X_8$: pupil-teacher ratio by town,
+
+$X_9$: $1,000(Bk-0.63)^2$ where $Bk$ is the proportion of people of Afrom American descent by town,
+
+$X_{10}$: percent lower status of population,
+
+and the $Y$ response variable is the median value of the owner-occupied homes in 1,000 USD. Then, the additive model assumed is
+$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
+
+
+```{r read the dataset}
+datos <- read.csv("housing.csv", header = FALSE, sep="")
+x1 <- datos$V1
+x2 <- datos$V3
+x3 <- datos$V5
+x4 <- datos$V6
+x5 <- datos$V7
+x6 <- datos$V8
+x7 <- datos$V10
+x8 <- datos$V11
+x9 <- datos$V12
+x10 <- datos$V13
+X <- log(cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10))
+colnames(X) <- NULL
+y <- datos$V14
 ```
 
[…]
