+++ vignettes/Examples.R
@@ -4,263 +4,35 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-## ----generating data----------------------------------------------------------
-function.g1 <- function(x1) 24*(x1-1/2)^2-2
-function.g2 <- function(x2) 2*pi*sin(pi*x2)-4
-set.seed(140)
-n <- 500
-x1 <- runif(n)
-x2 <- runif(n)
-X <- cbind(x1, x2)
-eps <- rnorm(n,0,sd=0.15)
-regression <- function.g1(x1)+function.g2(x2)
-y <- regression + eps
+## ----library------------------------------------------------------------------
+library("RBF")
 
-## ----bandw1-------------------------------------------------------------------
-bandw <- c(0.05, 0.075)
+## ----read the dataset---------------------------------------------------------
+datos <- read.csv("housing.csv", header = FALSE, sep="")
+x1 <- datos$V1
+x2 <- datos$V3
+x3 <- datos$V5
+x4 <- datos$V6
+x5 <- datos$V7
+x6 <- datos$V8
+x7 <- datos$V10
+x8 <- datos$V11
+x9 <- datos$V12
+x10 <- datos$V13
+X <- log(cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10))
+colnames(X) <- NULL
+y <- datos$V14
+
+## ----bandwidths---------------------------------------------------------------
+bandw <- 1/2*apply(X,2,sd)
+bandw
 
 ## ----robust fit---------------------------------------------------------------
-library(RBF)
-point <- c(0.7, 0.6)
-robust.fit <- backf.rob(y ~ X, point=point, windows=bandw, type = 'Tukey', degree=1)
+robust.fit <- backf.rob( y~X, windows=bandw)
 
-## ----prediction---------------------------------------------------------------
-robust.fit$prediction
-c(function.g1(point[1]), function.g2(point[2]))
+## ----summary------------------------------------------------------------------
+summary(robust.fit)
 
-## ----plots-simu, fig.show="hold", out.width="33%"-----------------------------
-lim.rob <- matrix(0, 2, 2)
-functions.g <- cbind(function.g1(X[,1]), function.g2(X[,2]))
-for(j in 1:2) {
-  res <- y - robust.fit$alpha - robust.fit$g.matrix[,-j]
-  lim.rob[,j] <- range(res)
-  plot(X[,j], res, type='p', pch=19, col='gray45', xlab=colnames(X)[j], ylab='', 
-       cex=1, ylim=lim.rob[,j])
-  ord <- order(X[,j])
-  lines(X[ord,j], robust.fit$g.matrix[ord,j], lwd=3, col='blue')
-  lines(X[ord,j], functions.g[ord,j], lwd=3)
-}
-
-## ----contaminated responses---------------------------------------------------
-eps <- rnorm(n,0,sd=0.15)
-prop.cont <- 0.10
-ou <- rbinom(n, size=1, prob=prop.cont) 
-eps.ou <- eps
-eps.ou[ ou == 1 ] <- rnorm(sum(ou),mean=15, sd=0.1)
-yout <- regression + eps.ou
-
-## ----bandw-out----------------------------------------------------------------
-bandw <- c(0.05, 0.075)
-
-## ----robust.fit.out-----------------------------------------------------------
[…]
