+++ vignettes/Examples.R
@@ -1,13 +1,13 @@
-## ----setup, include=FALSE-----------------------------------------------------
+## ----setup, include=FALSE----------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----library------------------------------------------------------------------
+## ----library-----------------------------------------------------
 library("RBF")
 
-## ----read the dataset---------------------------------------------------------
+## ----read the dataset--------------------------------------------
 datos <- read.csv("housing.csv", header = FALSE, sep="")
 x1 <- datos$V1
 x2 <- datos$V3
@@ -23,16 +23,16 @@ X <- log(cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10))
 colnames(X) <- NULL
 y <- datos$V14
 
-## ----bandwidths---------------------------------------------------------------
+## ----bandwidths--------------------------------------------------
 bandw <- 1/2*apply(X,2,sd)
 bandw
 
-## ----robust fit---------------------------------------------------------------
+## ----robust fit--------------------------------------------------
 robust.fit <- backf.rob( y~X, windows=bandw)
 
-## ----summary------------------------------------------------------------------
+## ----summary-----------------------------------------------------
 summary(robust.fit)
 
-## ----plot---------------------------------------------------------------------
+## ----plot--------------------------------------------------------
 plot(robust.fit)
 
