+++ vignettes/Examples.Rmd
@@ -70,24 +70,23 @@ $Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
 
 ```{r read the dataset}
 datos <- read.csv("housing.csv", header = FALSE, sep="")
-x1 <- datos$V1
-x2 <- datos$V3
-x3 <- datos$V5
-x4 <- datos$V6
-x5 <- datos$V7
-x6 <- datos$V8
-x7 <- datos$V10
-x8 <- datos$V11
-x9 <- datos$V12
-x10 <- datos$V13
-X <- log(cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10))
-colnames(X) <- NULL
+x1 <- datos$V1 #log(datos$V1)
+x2 <- datos$V3 #log(datos$V3)
+x3 <- datos$V5 #log(datos$V5)
+x4 <- datos$V6 #log(datos$V6)
+x5 <- datos$V7 #log(datos$V7)
+x6 <- datos$V8 #log(datos$V8)
+x7 <- datos$V10 #log(datos$V10)
+x8 <- datos$V11 # log(datos$V11)
+x9 <- datos$V12 #log(datos$V12)
+x10 <- datos$V13 #log(datos$V13)
 y <- datos$V14
 ```
 
 Recall that bandwidth selection is an important issue when considering any kernel-based method. As in Härdle et al. (2004), bandwidths a $h_j$, for $1 \leq j \leq 10$ are chosen proportional to its standard deviation ($h_j=0.5 \hat{\sigma}_j$).
 ```{r bandwidths}
-bandw <- 1/2*apply(X,2,sd)
+X <- log(cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10)) #log(cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10))
+bandw <- (1/2)*apply(X,2,sd) 
 bandw
 ```
 
@@ -105,8 +104,55 @@ summary(robust.fit)
 ```
 and we can obtain from the method `plot` the plots of the partial residuals together with the estimated curves. Once the fit is done and estimations of the constant $\mu$ and the additive functions $g_j$ are obtained, the partial residual associated to the  $j$th covariate for the $i$th observation is defined as $$R_{ij}=Y_i-\hat{\mu}-\sum_{k\neq i}\hat{g}_k(X_{ik}).$$ Then, points of the partial residuals obtained for the $j$th covariate are of the form $(X_{ij},R_{ij})$.
 ```{r plot}
-plot(robust.fit)
+plot(robust.fit, which=c(1,2,4,10))
 ```
 
 
+If you want a prediction of the additive functions $\hat{g}_j$ at a certain point, you can add it in the argument `point` when doing the fit.
+```{r prediction}
+po <- colMeans(X)
+robust.fit1 <- backf.rob( y~X, windows = bandw, point = po)
+robust.fit1$prediction
+```
 
+A PARTIR DE ACA LO QUE NO ME GUSTA ES QUE CON GAM LAS CURVAS NO DAR PARECIDAS, MIENTRAS QUE SI LAS DAN SI USAMOS EL ESTIMADOR CLASICO DE NUESTRO PAQUETE. HABRÁ MANERA DE USAR ESTE MISMO EN LUGAR DE USAR EL DE, POR EJEMPLO, GAM? PORQUE SI ES ASI, LO QUE SE PUEDE HACER ES CONTAMINAR UN PEQUEÑO PORCENTAJE DE OBSERVACIONES Y VER COMO SE MODIFICAN LAS CURVAS.
+
[…]
