+++ vignettes/Examples.R
@@ -1,13 +1,13 @@
-## ----setup, include=FALSE------------------------------------------------------------------------------------------------------------------
+## ----setup, include=FALSE-----------------------------------------------------------------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----library-------------------------------------------------------------------------------------------------------------------------------
+## ----library------------------------------------------------------------------------------------------------------------------------------
 library("RBF")
 
-## ----read the dataset----------------------------------------------------------------------------------------------------------------------
+## ----read the dataset---------------------------------------------------------------------------------------------------------------------
 datos <- read.csv("housing.csv", header = FALSE, sep="")
 x1 <- log(datos$V1)
 x2 <- log(datos$V3)
@@ -21,38 +21,38 @@ x9 <- log(datos$V12)
 x10 <- log(datos$V13)
 y <- datos$V14
 
-## ----bandwidths----------------------------------------------------------------------------------------------------------------------------
+## ----bandwidths---------------------------------------------------------------------------------------------------------------------------
 X <- cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10)
 colnames(X) <- NULL #HAY QUE ARREGLAR LOS NOMBRES CUANDO VIENEN CON NOMBRES EN EL .R DEL PAQUETE
 bandw <- (1/2)*apply(X,2,sd) 
 bandw
 
-## ----robust fit--------------------------------------------------
+## ----robust fit---------------------------------------------------------------------------------------------------------------------------
 robust.fit <- backf.rob(y~X, windows=bandw)
 
-## ----summary-----------------------------------------------------
+## ----summary------------------------------------------------------------------------------------------------------------------------------
 summary(robust.fit)
 
-## ----plot--------------------------------------------------------
+## ----plot---------------------------------------------------------------------------------------------------------------------------------
 plot(robust.fit)
 
-## ----prediction--------------------------------------------------
+## ----prediction---------------------------------------------------------------------------------------------------------------------------
 po <- colMeans(X)
 robust.fit1 <- backf.rob(y~X, windows = bandw, point = po)
 robust.fit1$prediction
 
-## ----outliers----------------------------------------------------
+## ----outliers-----------------------------------------------------------------------------------------------------------------------------
 ynew <- y
 #ynew[c(4,89,197,198,199,200,299,291,301,350)]<- rep(400,10)
 ynew[1:5]<- rep(400,5)
 
-## ----robustplotswithoutliers-------------------------------------
+## ----robustplotswithoutliers--------------------------------------------------------------------------------------------------------------
 robust.fit.new <- backf.rob(ynew~X, windows = bandw, point = po)
 summary(robust.fit.new)
 robust.fit.new$prediction
 plot(robust.fit.new)
 
-## ----robustplots2, warning=FALSE---------------------------------
+## ----robustplots2, warning=FALSE----------------------------------------------------------------------------------------------------------
 for(j in 1:10) {
   name.x <- bquote(paste('x')[.(j)]) 
   name.y <- bquote(paste(hat('g')[.(j)]))
@@ -62,7 +62,7 @@ for(j in 1:10) {
   lines(X[oo,j], robust.fit$g.matrix[oo,j], lwd=5, col='green', lty=2)
 }
 
-## ----gam, warning=FALSE------------------------------------------
+## ----gam, warning=FALSE-------------------------------------------------------------------------------------------------------------------
 library(gam)
[…]
