+++ doc/Examples.Rmd
@@ -1,0 +1,183 @@
+---
+title: "Examples"
+author: "Martínez and Salibian"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Examples}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+## About this vignette
+
+In this vignette, we will show with a real data example and some contaminated responses how a robust backfitting procedure can be easily implemented.
+
+The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF)) 
+implements the robust back-fitting algorithm as proposed by
+Boente, Martinez and Salibian-Barrera  in 
+
+> Boente G, Martinez A, Salibian-Barrera M. (2017) Robust estimators for 
+> additive models using backfitting. Journal of Nonparametric Statistics. Taylor 
+> & Francis; 29, 744-767.
+> [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077)
+
+This repository contains a development version of `RBF`
+which may differ slightly from the one available on CRAN
+(until the CRAN version is updated appropriately). 
+
+The package in this repository can be installed from within `R` by using the following code (assuming the [devtools](https://cran.r-project.org/package=devtools)) package is available:
+```R
+devtools::install_github("msalibian/RBF")
+```
+and charged in the `R` session by the following command
+```{r library}
+library("RBF")
+```
+Now that the `R` package is downloaded, we can now start to see how this procedure works.
+
+We consider the Boston house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). File `housing.csv` available in this GitHub repository contains $n=506$ observations for the census districts of the Boston metropolitan area. As in this book, we selected the following ten explanatory variables
+
+$X_1$: per capita crime rate by town,
+
+$X_2$: proportion of non-retail business acres per town,
+
+$X_3$: nitric oxides concentration (parts per 10 million),
+
+$X_4$: avarage number of rooms per dwelling,
+
+$X_5$: proportion of owner-occupied units built prior to 1940,
+
+$X_6$: weighted distances to five Boston employment centers,
+
+$X_7$: full-value property tax rate per $10,000,
+
+$X_8$: pupil-teacher ratio by town,
+
+$X_9$: $1,000(Bk-0.63)^2$ where $Bk$ is the proportion of people of Afrom American descent by town,
+
+$X_{10}$: percent lower status of population,
+
+and the $Y$ response variable is the median value of the owner-occupied homes in 1,000 USD. Then, the additive model assumed is
+$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
+
+
+```{r read the dataset}
+datos <- read.csv("housing.csv", header = FALSE, sep="")
+x1 <- log(datos$V1)
+x2 <- log(datos$V3)
+x3 <- log(datos$V5)
+x4 <- log(datos$V6)
+x5 <- log(datos$V7)
+x6 <- log(datos$V8)
+x7 <- log(datos$V10)
+x8 <- log(datos$V11)
+x9 <- log(datos$V12)
+x10 <- log(datos$V13)
[…]
