+++ vignettes/Examples.Rmd
@@ -17,97 +17,82 @@ knitr::opts_chunk$set(
 ```
 ## About this vignette
 
-In this vignette, we will show with a real data example and some contaminated responses how a robust backfitting procedure can be easily implemented.
+In this vignette we illustrate the use and properties of the
+robust backfitting estimator for additive models proposed in 
 
-The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF)) 
-implements the robust back-fitting algorithm as proposed by
-Boente, Martinez and Salibian-Barrera  in 
+> Boente G, Martinez A, Salibian-Barrera M. (2017) Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics. Taylor & Francis; 29, 744-767. [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077)
 
-> Boente G, Martinez A, Salibian-Barrera M. (2017) Robust estimators for 
-> additive models using backfitting. Journal of Nonparametric Statistics. Taylor 
-> & Francis; 29, 744-767.
-> [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077)
 
-This repository contains a development version of `RBF`
-which may differ slightly from the one available on CRAN
-(until the CRAN version is updated appropriately). 
+<!-- The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF))  -->
+<!-- implements the robust back-fitting algorithm as proposed by -->
+<!-- Boente, Martinez and Salibian-Barrera  in  -->
 
-The package in this repository can be installed from within `R` by using the following code (assuming the [devtools](https://cran.r-project.org/package=devtools)) package is available:
-```R
-devtools::install_github("msalibian/RBF")
-```
-and charged in the `R` session by the following command
-```{r library}
-library("RBF")
-```
-Now that the `R` package is downloaded, we can now start to see how this procedure works.
+<!-- This repository contains a development version of `RBF` -->
+<!-- which may differ slightly from the one available on CRAN -->
+<!-- (until the CRAN version is updated appropriately).  -->
 
-We consider the Boston house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). File `housing.csv` available in this GitHub repository contains $n=506$ observations for the census districts of the Boston metropolitan area. As in this book, we selected the following ten explanatory variables
+<!-- The package in this repository can be installed from within `R` by using the following code (assuming the [devtools](https://cran.r-project.org/package=devtools)) package is available: -->
+<!-- ```R -->
+<!-- devtools::install_github("msalibian/RBF") -->
+<!-- ``` -->
+<!-- and charged in the `R` session by the following command -->
+<!-- ```{r library} -->
+<!-- library("RBF") -->
+<!-- ``` -->
+<!-- Now that the `R` package is downloaded, we can now start to see how this procedure works. -->
 
-$X_1$: per capita crime rate by town,
[…]
