+++ doc/Examples.R
@@ -4,31 +4,20 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-## ----library------------------------------------------------------------------
-library("RBF")
-
 ## ----read the dataset---------------------------------------------------------
-datos <- read.csv("housing.csv", header = FALSE, sep="")
-x1 <- log(datos$V1)
-x2 <- log(datos$V3)
-x3 <- log(datos$V5)
-x4 <- log(datos$V6)
-x5 <- log(datos$V7)
-x6 <- log(datos$V8)
-x7 <- log(datos$V10)
-x8 <- log(datos$V11)
-x9 <- log(datos$V12)
-x10 <- log(datos$V13)
-y <- datos$V14
+data(Boston, package='MASS')
+dd <- Boston[, c(1, 3, 5:8, 10:14)]
+dd[, -11] <- log( dd[, names(dd) != 'medv'] )
+dd$medv <- log(dd$medv)
+
+## ----loadpckg-----------------------------------------------------------------
+library(RBF)
 
 ## ----bandwidths---------------------------------------------------------------
-X <- cbind(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10)
-colnames(X) <- NULL #HAY QUE ARREGLAR LOS NOMBRES CUANDO VIENEN CON NOMBRES EN EL .R DEL PAQUETE
-bandw <- (1/2)*apply(X,2,sd) 
-bandw
+bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 
-## ----robust fit---------------------------------------------------------------
-robust.fit <- backf.rob(y~X, windows=bandw)
+## ----robust fit, cache=TRUE---------------------------------------------------
+robust.fit <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw)
 
 ## ----summary------------------------------------------------------------------
 summary(robust.fit)
@@ -36,54 +25,65 @@ summary(robust.fit)
 ## ----plot---------------------------------------------------------------------
 plot(robust.fit)
 
-## ----prediction---------------------------------------------------------------
-po <- colMeans(X)
-robust.fit1 <- backf.rob(y~X, windows = bandw, point = po)
+## ----prediction, cache=TRUE---------------------------------------------------
+po <- colMeans(dd[, names(dd) != 'medv'])
+robust.fit1 <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw, point=po)
+
+## ----showpred-----------------------------------------------------------------
 robust.fit1$prediction
 
 ## ----outliers-----------------------------------------------------------------
-ynew <- y
-#ynew[c(4,89,197,198,199,200,299,291,301,350)]<- rep(400,10)
-ynew[1:5]<- rep(400,5)
+dd2 <- dd
+dd2$medv[1:5]<- rep(6, 5) # rep(400, 5)
 
-## ----robustplotswithoutliers--------------------------------------------------
-robust.fit.new <- backf.rob(ynew~X, windows = bandw, point = po)
+## ----robustplotswithoutliers, cache=TRUE--------------------------------------
+robust.fit.new <- backf.rob(medv ~ ., data = dd2, degree=0, type='Huber', windows=bandw, point = po)
 summary(robust.fit.new)
 robust.fit.new$prediction
 plot(robust.fit.new)
 
[…]
