+++ vignettes/Examples.Rmd
@@ -61,6 +61,11 @@ $Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
 dd[, -11] <- log( dd[, names(dd) != 'medv'] )
+dd$medv <- log(dd$medv)
+```
+
+```{r loadpckg}
+library(RBF)
 ```
 
 The backfitting estimators for each additive component are 
@@ -80,7 +85,7 @@ the robust backfitting estimator are:
 - `degree`: the degree of the  polynomial used for the kernel local regression, defaults to `0`,
 - `type`: specifies the robust loss function, options are `Huber` or `Tukey`, defaults to `Huber`.
 The command to compute the fit is:
-```{r robust fit}
+```{r robust fit, cache=TRUE}
 robust.fit <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw)
 ```
 The summary method provides information on the resulting fit: 
@@ -94,22 +99,32 @@ residuals: $$R_{ij}=Y_i-\hat{\mu}-\sum_{k\neq i}\hat{g}_k(X_{ik})$$.
 plot(robust.fit)
 ```
 
-If you want a prediction of the additive functions $\hat{g}_j$ at a certain point, you can add it in the argument `point` when doing the fit.
-```{r prediction}
-po <- colMeans(X)
-robust.fit1 <- backf.rob(y~X, windows = bandw, point = po)
+By default, `backf.rob` computes fitted values on the training
+set. If predictions at a different specific point are desired, 
+we can pass those points using the argumen `point`. For example, 
+to obtain predicted values at the point given by average of 
+each explanatory variable, we can use the following 
+command (note that this step implies re-fitting the whole model):
+```{r prediction, cache=TRUE}
+po <- colMeans(dd[, names(dd) != 'medv'])
+robust.fit1 <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw, point=po)
+```
+The values of the estimated components evaluated at the 
+corresponding coordinates of `po` are 
+returned in the `$prediction` element: 
+```{r showpred}
 robust.fit1$prediction
 ```
-
-In order to analyze the sensitivity of the robust estimator, we will contaminate the responses of a 1% of the observations.
+In order to illustrate the behaviour of the robust fit when 
+outliers are present in the data, we artifically introduce
+1% of atypical values in the response variable: 
 ```{r outliers}
-ynew <- y
-#ynew[c(4,89,197,198,199,200,299,291,301,350)]<- rep(400,10)
-ynew[1:5]<- rep(400,5)
+dd2 <- dd
+dd2$medv[1:5]<- rep(6, 5) # rep(400, 5)
 ```
 Then, if the recalculate the robust fit, we obtain similar curves to what was obtained with the original data set.
-```{r robustplotswithoutliers}
-robust.fit.new <- backf.rob(ynew~X, windows = bandw, point = po)
+```{r robustplotswithoutliers, cache=TRUE}
+robust.fit.new <- backf.rob(medv ~ ., data = dd2, degree=0, type='Huber', windows=bandw, point = po)
 summary(robust.fit.new)
[…]

[…]
