+++ doc/Examples.Rmd
@@ -61,7 +61,6 @@ $Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
 dd[, -11] <- log( dd[, names(dd) != 'medv'] )
-dd$medv <- log(dd$medv)
 ```
 
 ```{r loadpckg}
