+++ vignettes/Examples.Rmd
@@ -41,6 +41,8 @@ robust backfitting estimator for additive models proposed in
 <!-- ``` -->
 <!-- Now that the `R` package is downloaded, we can now start to see how this procedure works. -->
 
+### Boston example
+
 We will use the well-known `Boston` house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). The data are available in the `MASS` package. It contains $n=506$ observations and 14 variables measured on the census districts of the Boston metropolitan area. Following the analysis in Härdle et a. (2004) we use the following 10 explanatory variables:
 
 - `crim`: per capita crime rate by town ($X_1$),
@@ -49,7 +51,7 @@ We will use the well-known `Boston` house price data of Harrinson and Rubinfeld 
 - `rm`: average number of rooms per dwelling ($X_4$),
 - `age`: proportion of owner-occupied units built prior to 1940 ($X_5$),
 - `dis`: weighted distances to five Boston employment centers ($X_6$),
-- `tax`: full-value property tax rate per $10,000 ($X_7$),
+- `tax`: full-value property tax rate per 10,000 ($X_7$),
 - `ptratio`: pupil-teacher ratio by town ($X_8$),
 - `black`: $1,000(Bk-0.63)^2$ where $Bk$ is the proportion of people of Afrom American descent by town ($X_9$),
 - `lstat`: percent lower status of population ($X_{10}$),
@@ -73,16 +75,18 @@ bandwidth selection is an important step. We follow
 Härdle et al. (2004) and select bandwidths $h_j$, $1 \le j \le 10$, 
 proportional to the standard deviation of the corresponding
 explanatory variables $X_j$, specifically we set 
-$h_j = \hat{\sigma}_j / 2$):
+$h_j = \hat{\sigma}_j / 2$: OBS! EN REALIDAD ES EL DESVIO DEL LOGARITMO DE LAS $X_j$.
 ```{r bandwidths}
 bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 ```
 To compute the robust backfitting estimator we specify the model using the
 standard `formula` notation in `R`. Other necessary arguments to compute
 the robust backfitting estimator are:
+
 - `windows`: the bandwidths,
 - `degree`: the degree of the  polynomial used for the kernel local regression, defaults to `0`,
 - `type`: specifies the robust loss function, options are `Huber` or `Tukey`, defaults to `Huber`.
+
 The command to compute the fit is:
 ```{r robust fit, cache=TRUE}
 robust.fit <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw)
@@ -92,8 +96,8 @@ The summary method provides information on the resulting fit:
 summary(robust.fit)
 ```
 while the plot method can be used to visualize the estimated 
-additive components (displayed over the corresponding partial
-residuals: $$R_{ij}=Y_i-\hat{\mu}-\sum_{k\neq i}\hat{g}_k(X_{ik})$$. 
+additive components displayed over the corresponding partial
+residuals: $$R_{ij}=Y_i-\hat{\mu}-\sum_{k\neq i}\hat{g}_k(X_{ik}).$$ 
 ```{r plot}
 plot(robust.fit)
 ```
@@ -182,9 +186,345 @@ for(j in 1:10) {
 }
 ```
 
[…]
