### 2021-03-02 590b2fb
update vignette, eval = FALSE, etc.
