+++ doc/Examples.R
@@ -86,3 +86,251 @@ for(j in 1:10) {
   lines(dd2[oo,j], fits[oo,j], lwd=5, col='darkorange2', lty=2)
 }
 
+## ----scatterplot--------------------------------------------------------------
+data(airquality)
+ccs <- complete.cases(airquality)
+aircomplete <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
+# pairs(aircomplete[, c('Ozone', 'Solar.R', 'Wind', 'Temp')], pch=19, col='gray30', cex=1.5)
+
+## ----robustcv, warning=FALSE, cache=TRUE, eval=FALSE--------------------------
+#  library(RBF)
+#  
+#  # Bandwidth selection with leave-one-out cross-validation
+#  ## Without outliers
+#  # This takes a long time to compute (approx 380 minutes running
+#  # R 3.6.1 on an Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz)
+#  a <- c(1/2, 1, 1.5, 2, 2.5, 3)
+#  h1 <- a * sd(aircomplete[,2])
+#  h2 <- a * sd(aircomplete[,3])
+#  h3 <- a * sd(aircomplete[,4])
+#  hh <- expand.grid(h1, h2, h3)
+#  nh <- nrow(hh)
+#  rmspe <- rep(NA, nh)
+#  jbest <- 0
+#  cvbest <- +Inf
+#  n <- nrow(aircomplete)
+#  for(i in 1:nh) {
+#    # leave-one-out CV loop
+#    preds <- rep(NA, n)
+#    for(j in 1:n) {
+#      tmp <- try( backf.rob(Ozone ~ Solar.R + Wind + Temp, point = aircomplete[j, -1],
+#                            windows = hh[i, ], epsilon = 1e-6, data=aircomplete,
+#                            degree = 1, type = 'Tukey', subset = c(-j) ))
+#      if (class(tmp)[1] != "try-error") {
+#        preds[j] <- rowSums(tmp$prediction) + tmp$alpha
+#      }
+#    }
+#    tmp.re <- RobStatTM::locScaleM(preds - aircomplete$Ozone, na.rm=TRUE)
+#    rmspe[i] <- tmp.re$mu^2 + tmp.re$disper^2
+#    if( rmspe[i] < cvbest ) {
+#      jbest <- i
+#      cvbest <- rmspe[i]
+#    }
+#  }
+#  (bandw <- hh[jbest,])
+
+## ----bandw--------------------------------------------------------------------
+bandw <- c(136.7285, 10.67314, 4.764985)
+
+## ----fitfull------------------------------------------------------------------
+fit.full <- backf.rob(Ozone ~ Solar.R + Wind + Temp, windows=bandw, 
+                      epsilon=1e-6, degree=1, type='Tukey', 
+                      subset = ccs, data=airquality)
+
+## ----plotfitfull--------------------------------------------------------------
+plot(fit.full)
+
+## ----gamcv, cache=TRUE, warning=FALSE, eval=FALSE-----------------------------
+#  library(gam)
+#  a <- c(.3, .4, .5, .6, .7, .8, .9)
+#  hh <- expand.grid(a, a, a)
+#  nh <- nrow(hh)
+#  jbest <- 0
+#  cvbest <- +Inf
+#  n <- nrow(aircomplete)
+#  for(i in 1:nh) {
+#    fi <- rep(0, n)
+#    for(j in 1:n) {
+#      tmp <- gam(Ozone ~ lo(Solar.R, span=hh[i,1]) + lo(Wind, span=hh[i,2])
[…]
