+++ vignettes/Examples.Rmd
@@ -2,7 +2,7 @@
 title: "Examples"
 author: "Martínez and Salibian"
 date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
+output: pdf_document
 vignette: >
   %\VignetteIndexEntry{Examples}
   %\VignetteEngine{knitr::rmarkdown}
@@ -57,7 +57,7 @@ We will use the well-known `Boston` house price data of Harrinson and Rubinfeld 
 - `lstat`: percent lower status of population ($X_{10}$),
 
 The response variable $Y$ is `medv`, the median value of the owner-occupied homes in 1,000 USD, and the proposed  additive model is
-$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
+$$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$$
 
 ```{r read the dataset}
 data(Boston, package='MASS')
@@ -75,7 +75,7 @@ bandwidth selection is an important step. We follow
 Härdle et al. (2004) and select bandwidths $h_j$, $1 \le j \le 10$, 
 proportional to the standard deviation of the corresponding
 explanatory variables $X_j$, specifically we set 
-$h_j = \hat{\sigma}_j / 2$: OBS! EN REALIDAD ES EL DESVIO DEL LOGARITMO DE LAS $X_j$.
+$h_j = \hat{\sigma}_j / 2$: \textcolor{red}{OBS! EN REALIDAD ES EL DESVIO DEL LOGARITMO DE LAS $X_j$.}
 ```{r bandwidths}
 bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 ```
@@ -89,7 +89,8 @@ the robust backfitting estimator are:
 
 The command to compute the fit is:
 ```{r robust fit, cache=TRUE}
-robust.fit <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw)
+robust.fit <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
+                        windows = bandw)
 ```
 The summary method provides information on the resulting fit: 
 ```{r summary}
@@ -110,7 +111,8 @@ each explanatory variable, we can use the following
 command (note that this step implies re-fitting the whole model):
 ```{r prediction, cache=TRUE}
 po <- colMeans(dd[, names(dd) != 'medv'])
-robust.fit1 <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw, point=po)
+robust.fit1 <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
+                         windows = bandw, point = po)
 ```
 The values of the estimated components evaluated at the 
 corresponding coordinates of `po` are 
@@ -230,7 +232,7 @@ for(i in 1:nh) {
   preds <- rep(NA, n)
   for(j in 1:n) {
     tmp <- try( backf.rob(Ozone ~ Solar.R + Wind + Temp, point = aircomplete[j, -1],
-                          windows = hh[i, ], epsilon = 1e-6, data=aircomplete,
+                          windows = hh[i, ], epsilon = 1e-6, data = aircomplete,
                           degree = 1, type = 'Tukey', subset = c(-j) ))
     if (class(tmp)[1] != "try-error") {
       preds[j] <- rowSums(tmp$prediction) + tmp$alpha
@@ -252,9 +254,9 @@ bandw <- c(136.7285, 10.67314, 4.764985)
 Now we use the robust backfitting algorithm to fit an additive model using Tukey's bisquare loss (the default tuning constant for this loss function is 4.685) and the optimal bandwidths.
 
 ```{r fitfull}
[…]
