+++ vignettes/Examples.R
@@ -1,49 +1,34 @@
-## ----setup, include=FALSE-----------------------------------------------------
+## ----setup, include=FALSE------------------------------------------------------------------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----read the dataset---------------------------------------------------------
+## ----read the dataset----------------------------------------------------------------------------------------------------------------------
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
 dd[, -11] <- log( dd[, names(dd) != 'medv'] )
 
-## ----loadpckg-----------------------------------------------------------------
+## ----loadpckg------------------------------------------------------------------------------------------------------------------------------
 library(RBF)
 
-## ----bandwidths---------------------------------------------------------------
+## ----bandwidths----------------------------------------------------------------------------------------------------------------------------
 bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 
-## ----robustfit, cache=TRUE----------------------------------------------------
-robust.fit <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
-                        windows = bandw)
-
-## ----summary------------------------------------------------------------------
+## ----summary-------------------------------------------------------------------------------------------------------------------------------
 summary(robust.fit)
 
-## ----plot---------------------------------------------------------------------
+## ----plot----------------------------------------------------------------------------------------------------------------------------------
 plot(robust.fit)
 
-## ----preds, cache=TRUE--------------------------------------------------------
-po <- colMeans(dd[, names(dd) != 'medv'])
-robust.fit1 <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
-                         windows = bandw, point = po)
-
-## ----showpred-----------------------------------------------------------------
+## ----showpred------------------------------------------------------------------------------------------------------------------------------
 robust.fit1$prediction
 
-## ----outliers-----------------------------------------------------------------
+## ----outliers------------------------------------------------------------------------------------------------------------------------------
 dd2 <- dd
 dd2$medv[1:5]<- rep(400, 5)
 
-## ----robustplotswithoutliers, cache=TRUE--------------------------------------
-robust.fit.new <- backf.rob(medv ~ ., data = dd2, degree=0, type='Huber', windows=bandw, point = po)
-summary(robust.fit.new)
-robust.fit.new$prediction
-plot(robust.fit.new)
-
-## ----robustplots2, warning=FALSE----------------------------------------------
+## ----robustplots2, warning=FALSE-----------------------------------------------------------------------------------------------------------
 for(j in 1:10) {
   name.x <- names(dd)[j] 
   name.y <- bquote(paste(hat('g')[.(j)]))
[…]
