+++ vignettes/Examples.R
@@ -1,34 +1,34 @@
-## ----setup, include=FALSE------------------------------------------------------------------------------------------------------------------
+## ----setup, include=FALSE-----------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----read the dataset----------------------------------------------------------------------------------------------------------------------
+## ----read the dataset---------------------------------------------------------
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
 dd[, -11] <- log( dd[, names(dd) != 'medv'] )
 
-## ----loadpckg------------------------------------------------------------------------------------------------------------------------------
+## ----loadpckg-----------------------------------------------------------------
 library(RBF)
 
-## ----bandwidths----------------------------------------------------------------------------------------------------------------------------
+## ----bandwidths---------------------------------------------------------------
 bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 
-## ----summary-------------------------------------------------------------------------------------------------------------------------------
+## ----summary------------------------------------------------------------------
 summary(robust.fit)
 
-## ----plot----------------------------------------------------------------------------------------------------------------------------------
+## ----plot---------------------------------------------------------------------
 plot(robust.fit)
 
-## ----showpred------------------------------------------------------------------------------------------------------------------------------
+## ----showpred-----------------------------------------------------------------
 robust.fit1$prediction
 
-## ----outliers------------------------------------------------------------------------------------------------------------------------------
+## ----outliers-----------------------------------------------------------------
 dd2 <- dd
 dd2$medv[1:5]<- rep(400, 5)
 
-## ----robustplots2, warning=FALSE-----------------------------------------------------------------------------------------------------------
+## ----robustplots2, warning=FALSE----------------------------------------------
 for(j in 1:10) {
   name.x <- names(dd)[j] 
   name.y <- bquote(paste(hat('g')[.(j)]))
@@ -38,7 +38,7 @@ for(j in 1:10) {
   lines(dd2[oo,j], robust.fit$g.matrix[oo,j], lwd=5, col='green', lty=2)
 }
 
-## ----gam, warning=FALSE--------------------------------------------------------------------------------------------------------------------
+## ----gam, warning=FALSE-------------------------------------------------------
 library(gam)
 fit.gam <- gam(medv ~ lo(crim, span=1.62) + 
                  lo(indus, span=0.58) + 
@@ -63,7 +63,7 @@ fit.gam.new <- gam(medv ~ lo(crim, span=1.62) +
                  lo(lstat, span=0.45), data=dd2)
 fits.new <- predict(fit.gam.new, type='terms')
 
-## ----gamplots------------------------------------------------------------------------------------------------------------------------------
+## ----gamplots-----------------------------------------------------------------
 for(j in 1:10) {
   oo <- order(dd2[,j])
   name.x <- names(dd)[j] 
@@ -73,28 +73,28 @@ for(j in 1:10) {
   lines(dd2[oo,j], fits[oo,j], lwd=5, col='darkorange2', lty=2)
 }
 
-## ----scatterplot---------------------------------------------------------------------------------------------------------------------------
+## ----scatterplot--------------------------------------------------------------
 data(airquality)
 ccs <- complete.cases(airquality)
[…]
