+++ doc/Examples.R
@@ -15,8 +15,9 @@ library(RBF)
 ## ----bandwidths---------------------------------------------------------------
 bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 
-## ----robust fit, cache=TRUE---------------------------------------------------
-robust.fit <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw)
+## ----robustfit, cache=TRUE----------------------------------------------------
+robust.fit <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
+                        windows = bandw)
 
 ## ----summary------------------------------------------------------------------
 summary(robust.fit)
@@ -24,9 +25,10 @@ summary(robust.fit)
 ## ----plot---------------------------------------------------------------------
 plot(robust.fit)
 
-## ----prediction, cache=TRUE---------------------------------------------------
+## ----preds, cache=TRUE--------------------------------------------------------
 po <- colMeans(dd[, names(dd) != 'medv'])
-robust.fit1 <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw, point=po)
+robust.fit1 <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
+                         windows = bandw, point = po)
 
 ## ----showpred-----------------------------------------------------------------
 robust.fit1$prediction
@@ -90,7 +92,7 @@ for(j in 1:10) {
 data(airquality)
 ccs <- complete.cases(airquality)
 aircomplete <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
-# pairs(aircomplete[, c('Ozone', 'Solar.R', 'Wind', 'Temp')], pch=19, col='gray30', cex=1.5)
+pairs(aircomplete[, c('Ozone', 'Solar.R', 'Wind', 'Temp')], pch=19, col='gray30', cex=1.5)
 
 ## ----robustcv, warning=FALSE, cache=TRUE, eval=FALSE--------------------------
 #  library(RBF)
@@ -114,7 +116,7 @@ aircomplete <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
 #    preds <- rep(NA, n)
 #    for(j in 1:n) {
 #      tmp <- try( backf.rob(Ozone ~ Solar.R + Wind + Temp, point = aircomplete[j, -1],
-#                            windows = hh[i, ], epsilon = 1e-6, data=aircomplete,
+#                            windows = hh[i, ], epsilon = 1e-6, data = aircomplete,
 #                            degree = 1, type = 'Tukey', subset = c(-j) ))
 #      if (class(tmp)[1] != "try-error") {
 #        preds[j] <- rowSums(tmp$prediction) + tmp$alpha
@@ -133,9 +135,9 @@ aircomplete <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
 bandw <- c(136.7285, 10.67314, 4.764985)
 
 ## ----fitfull------------------------------------------------------------------
-fit.full <- backf.rob(Ozone ~ Solar.R + Wind + Temp, windows=bandw, 
-                      epsilon=1e-6, degree=1, type='Tukey', 
-                      subset = ccs, data=airquality)
[…]
