+++ doc/Examples.Rmd
@@ -57,7 +57,7 @@ We will use the well-known `Boston` house price data of Harrinson and Rubinfeld 
 - `lstat`: percent lower status of population ($X_{10}$),
 
 The response variable $Y$ is `medv`, the median value of the owner-occupied homes in 1,000 USD, and the proposed  additive model is
-$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$
+$$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$$
 
 ```{r read the dataset}
 data(Boston, package='MASS')
@@ -75,7 +75,7 @@ bandwidth selection is an important step. We follow
 Härdle et al. (2004) and select bandwidths $h_j$, $1 \le j \le 10$, 
 proportional to the standard deviation of the corresponding
 explanatory variables $X_j$, specifically we set 
-$h_j = \hat{\sigma}_j / 2$: OBS! EN REALIDAD ES EL DESVIO DEL LOGARITMO DE LAS $X_j$.
+$h_j = \hat{\sigma}_j / 2$: \textcolor{red}{OBS! EN REALIDAD ES EL DESVIO DEL LOGARITMO DE LAS $X_j$.}
 ```{r bandwidths}
 bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
 ```
@@ -88,8 +88,9 @@ the robust backfitting estimator are:
 - `type`: specifies the robust loss function, options are `Huber` or `Tukey`, defaults to `Huber`.
 
 The command to compute the fit is:
-```{r robust fit, cache=TRUE}
-robust.fit <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw)
+```{r robustfit, cache=TRUE}
+robust.fit <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
+                        windows = bandw)
 ```
 The summary method provides information on the resulting fit: 
 ```{r summary}
@@ -108,9 +109,10 @@ we can pass those points using the argumen `point`. For example,
 to obtain predicted values at the point given by average of 
 each explanatory variable, we can use the following 
 command (note that this step implies re-fitting the whole model):
-```{r prediction, cache=TRUE}
+```{r preds, cache=TRUE}
 po <- colMeans(dd[, names(dd) != 'medv'])
-robust.fit1 <- backf.rob(medv ~ ., data = dd, degree=0, type='Huber', windows=bandw, point=po)
+robust.fit1 <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber', 
+                         windows = bandw, point = po)
 ```
 The values of the estimated components evaluated at the 
 corresponding coordinates of `po` are 
@@ -201,8 +203,10 @@ The scatter plot suggests that the relationship between ozone and the other vari
 data(airquality)
 ccs <- complete.cases(airquality)
 aircomplete <- airquality[ccs, c('Ozone', 'Solar.R', 'Wind', 'Temp')]
-# pairs(aircomplete[, c('Ozone', 'Solar.R', 'Wind', 'Temp')], pch=19, col='gray30', cex=1.5)
+pairs(aircomplete[, c('Ozone', 'Solar.R', 'Wind', 'Temp')], pch=19, col='gray30', cex=1.5)
 ```
+
+\textcolor{red}{El pairs anterior estaba comentado. Si no lo vamos a poner, habria que sacar la parte del The scatter plot suggests me parece.}
 
[…]
