### 2021-03-03 b31e357
some editions, fixed sizes of figures
