+++ doc/Examples.R
@@ -7,7 +7,7 @@ knitr::opts_chunk$set(
 ## ----read the dataset---------------------------------------------------------
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
-dd[, -11] <- log( dd[, names(dd) != 'medv'] )
+dd[, names(dd) != 'medv'] <- log( dd[, names(dd) != 'medv'] )
 
 ## ----loadpckg-----------------------------------------------------------------
 library(RBF)
@@ -22,7 +22,7 @@ robust.fit <- backf.rob(medv ~ ., data = dd, degree = 0, type = 'Huber',
 ## ----summary------------------------------------------------------------------
 summary(robust.fit)
 
-## ----plot---------------------------------------------------------------------
+## ----plot, out.width  = "45%"-------------------------------------------------
 plot(robust.fit)
 
 ## ----preds, cache=TRUE--------------------------------------------------------
@@ -41,16 +41,15 @@ dd2$medv[1:5]<- rep(400, 5)
 robust.fit.new <- backf.rob(medv ~ ., data = dd2, degree=0, type='Huber', windows=bandw, point = po)
 summary(robust.fit.new)
 robust.fit.new$prediction
-plot(robust.fit.new)
 
-## ----robustplots2, warning=FALSE----------------------------------------------
+## ----robustplots2, warning=FALSE, out.width  = "45%"--------------------------
 for(j in 1:10) {
   name.x <- names(dd)[j] 
   name.y <- bquote(paste(hat('g')[.(j)]))
   oo <- order(dd2[,j])
-  plot(dd2[oo,j], robust.fit.new$g.matrix[oo,j], type="l", lwd=5, col='blue', lty=1, 
+  plot(dd2[oo,j], robust.fit.new$g.matrix[oo,j], type="l", lwd=2, col='blue', lty=1, 
        xlab=name.x, ylab=name.y)
-  lines(dd2[oo,j], robust.fit$g.matrix[oo,j], lwd=5, col='green', lty=2)
+  lines(dd2[oo,j], robust.fit$g.matrix[oo,j], lwd=2, col='green', lty=2)
 }
 
 ## ----gam, warning=FALSE-------------------------------------------------------
@@ -78,21 +77,21 @@ fit.gam.new <- gam(medv ~ lo(crim, span=1.62) +
                  lo(lstat, span=0.45), data=dd2)
 fits.new <- predict(fit.gam.new, type='terms')
 
-## ----gamplots-----------------------------------------------------------------
+## ----gamplots, out.width  = "45%"---------------------------------------------
 for(j in 1:10) {
   oo <- order(dd2[,j])
   name.x <- names(dd)[j] 
   name.y <- bquote(paste(hat('g')[.(j)]))
-  plot(dd2[oo,j], fits.new[oo,j], type="l", lwd=5, col='purple', lty=1, 
+  plot(dd2[oo,j], fits.new[oo,j], type="l", lwd=2, col='purple', lty=1, 
        xlab=name.x, ylab=name.y)
-  lines(dd2[oo,j], fits[oo,j], lwd=5, col='darkorange2', lty=2)
+  lines(dd2[oo,j], fits[oo,j], lwd=2, col='darkorange2', lty=2)
 }
 
-## ----scatterplot--------------------------------------------------------------
+## ----scatterplot, out.width  = "75%"------------------------------------------
 data(airquality)
 ccs <- complete.cases(airquality)
[…]
