+++ vignettes/Examples.Rmd
@@ -15,12 +15,25 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 ```
-## About this vignette
+# About this vignette
 
 In this vignette we illustrate the use and properties of the
-robust backfitting estimator for additive models proposed in 
+robust backfitting estimator for additive models implemented
+in the package `RBF`. This estimator was originally proposed in 
+Boente G, Martinez A, Salibian-Barrera M. (2017). 
 
-> Boente G, Martinez A, Salibian-Barrera M. (2017) Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics. Taylor & Francis; 29, 744-767. [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077)
+We will use two data sets. With the first one we illustrate
+the robustness of the robust backfitting estimators 
+when a small 
+proportion of very large outliers are present in the training
+set, and compare them with those obtained with the standard backfitting 
+algorithm. With the second example, we show how these robust estimators
+can be interpreted as automatically detecting and downweighting 
+potential outliers in the training set. We also compare the 
+prediction accuracy of the robust and classical backfitting 
+algorithms. 
+
+<!-- Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics. Taylor & Francis; 29, 744-767. [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077) -->
 
 
 <!-- The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF))  -->
@@ -41,7 +54,7 @@ robust backfitting estimator for additive models proposed in
 <!-- ``` -->
 <!-- Now that the `R` package is downloaded, we can now start to see how this procedure works. -->
 
-### Boston example
+## Boston example
 
 We will use the well-known `Boston` house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). The data are available in the `MASS` package. It contains $n=506$ observations and 14 variables measured on the census districts of the Boston metropolitan area. Following the analysis in Härdle et a. (2004) we use the following 10 explanatory variables:
 
@@ -62,7 +75,7 @@ $$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$$
 ```{r read the dataset}
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
-dd[, -11] <- log( dd[, names(dd) != 'medv'] )
+dd[, names(dd) != 'medv'] <- log( dd[, names(dd) != 'medv'] )
 ```
 
 ```{r loadpckg}
@@ -70,23 +83,24 @@ library(RBF)
 ```
 
 The backfitting estimators for each additive component are 
-computed using kernel-based local polynomial regression, and thus 
-bandwidth selection is an important step. We follow 
-Härdle et al. (2004) and select bandwidths $h_j$, $1 \le j \le 10$, 
-proportional to the standard deviation of the corresponding
-explanatory variables $X_j$, specifically we set 
-$h_j = \hat{\sigma}_j / 2$: \textcolor{red}{OBS! EN REALIDAD ES EL DESVIO DEL LOGARITMO DE LAS $X_j$.}
-```{r bandwidths}
-bandw <- apply(dd[, names(dd) != 'medv'], 2, sd) / 2
-```
[…]
