+++ vignettes/Examples.Rmd
@@ -17,17 +17,18 @@ knitr::opts_chunk$set(
 ```
 # About this vignette
 
-In this vignette we illustrate the use and properties of the
-robust backfitting estimator for additive models implemented
-in the package `RBF`. This estimator was originally proposed in 
+In this vignette we discuss some properties of a
+robust backfitting estimator for additive models, and 
+illustrate the use of the package `RBF` that implements it. 
+These estimators were originally proposed in 
 Boente G, Martinez A, Salibian-Barrera M. (2017). 
 
-We will use two data sets. With the first one we illustrate
-the robustness of the robust backfitting estimators 
+Below we analyze two data sets. The first one
+shows the robustness of the robust backfitting estimators 
 when a small 
 proportion of very large outliers are present in the training
-set, and compare them with those obtained with the standard backfitting 
-algorithm. With the second example, we show how these robust estimators
+set, and compares them with those obtained with the standard backfitting 
+algorithm. With the second example, we illustrate how these robust estimators
 can be interpreted as automatically detecting and downweighting 
 potential outliers in the training set. We also compare the 
 prediction accuracy of the robust and classical backfitting 
@@ -56,7 +57,7 @@ algorithms.
 
 ## Boston example
 
-We will use the well-known `Boston` house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). The data are available in the `MASS` package. It contains $n=506$ observations and 14 variables measured on the census districts of the Boston metropolitan area. Following the analysis in Härdle et a. (2004) we use the following 10 explanatory variables:
+Consider the well-known `Boston` house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). The data are available in the `MASS` package. It contains $n=506$ observations and 14 variables measured on the census districts of the Boston metropolitan area. Following the analysis in Härdle et a. (2004) we use the following 10 explanatory variables:
 
 - `crim`: per capita crime rate by town ($X_1$),
 - `indus`: proportion of non-retail business acres per town ($X_2$),
@@ -71,23 +72,27 @@ We will use the well-known `Boston` house price data of Harrinson and Rubinfeld 
 
 The response variable $Y$ is `medv`, the median value of the owner-occupied homes in 1,000 USD, and the proposed  additive model is
 $$Y= \mu+ \sum_{j=1}^{10} g_j(\log(X_j))+ \epsilon.$$
+where $\mu \in \mathbb{R}$, $g_j$ are unknown smooth functions,
+and $\epsilon$ are random errors. 
 
+First we load the data, and transform the explanatory variables as
+required by the model above:
 ```{r read the dataset}
 data(Boston, package='MASS')
 dd <- Boston[, c(1, 3, 5:8, 10:14)]
 dd[, names(dd) != 'medv'] <- log( dd[, names(dd) != 'medv'] )
 ```
-
+Next, we load the `RBF` package
 ```{r loadpckg}
 library(RBF)
 ```
-
-The backfitting estimators for each additive component are 
+The robust backfitting estimators for each additive component are 
 computed using robust 
 kernel-based local polynomial regression.
 The model to be fit is specified using the
-standard `formula` notation in `R`. Other necessary arguments to compute
[…]
