### 2021-03-05 #2992 @alemermartinez
Dear Mikkel,
We have just finished the vignette, which can be found in the "doc" folder of the main branch of the repository. I believe our revision is now complete.
Kind regards,
Alejandra

________________________________
De: Mikkel Meyer Andersen <notifications@github.com>
Enviado: jueves, 4 de marzo de 2021 18:41
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2992)

@alemermartinez<https://github.com/alemermartinez>, can you please give a brief status? This is not to rush you, merely to give me an impression of the progress and time-frame.
