### 2021-03-05 #2992 @mmrabe
@alemermartinez , the typical location for vignettes is `vignettes/`. The contents of `docs/` will not be visible to the user of the final package.
