### 2021-03-05 #2992 @alemermartinez
Thank you, Maximilian. We didn't know that. I copied Examples.html and Examples.R to "vignettes".

Kind regards,

Alejandra

________________________________
De: Maximilian M. Rabe <notifications@github.com>
Enviado: viernes, 5 de marzo de 2021 03:07
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Alejandra Martinez <ale_m_martinez@hotmail.com>; Mention <mention@noreply.github.com>
Asunto: Re: [openjournals/joss-reviews] [REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2992)

@alemermartinez<https://github.com/alemermartinez> , the typical location for vignettes is vignettes/. The contents of docs/ will not be visible to the user of the final package.
