### 2021-03-05 #2992 @msalibian
@mmrabe @alemermartinez perhaps the easiest way to access the vignette before the package is updated on CRAN would be to install it from the repository with
```
devtools::install_github('msalibian/RBF', build_vignettes=TRUE)
```
and then open it with
```
vignette('Examples', package='RBF')
```
