### 2021-03-19 #2992 @mikldk
@alemermartinez Please also read <https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Writing-package-vignettes> and <https://r-pkgs.org/vignettes.html>. Normally, `.R` and `.html` files are not in the `vignette` directory. Use e.g. `devtools::build_vignettes()` for handling vignettes (see [here](https://devtools.r-lib.org/reference/build_vignettes.html)).
