+++ doc/Examples.Rmd
@@ -1,574 +1,0 @@
----
-title: "Examples"
-author: "Martínez and Salibian"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Examples}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-# About this vignette
-
-In this vignette we discuss some properties of a
-robust backfitting estimator for additive models, and 
-illustrate the use of the package `RBF` that implements it. 
-These estimators were originally proposed in 
-Boente G, Martinez A, Salibian-Barrera M. (2017). 
-
-Below we analyze two data sets. The first one
-shows the robustness of the robust backfitting estimators 
-when a small 
-proportion of very large outliers are present in the training
-set, and compares them with those obtained with the standard backfitting 
-algorithm. With the second example, we illustrate how these robust estimators
-can be interpreted as automatically detecting and downweighting 
-potential outliers in the training set. We also compare the 
-prediction accuracy of the robust and classical backfitting 
-algorithms. 
-
-<!-- Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics. Taylor & Francis; 29, 744-767. [DOI: 10.1080/10485252.2017.1369077](https://doi.org/10.1080/10485252.2017.1369077) -->
-
-
-<!-- The `R` package `RBF` (available on CRAN [here](https://cran.r-project.org/package=RBF))  -->
-<!-- implements the robust back-fitting algorithm as proposed by -->
-<!-- Boente, Martinez and Salibian-Barrera  in  -->
-
-<!-- This repository contains a development version of `RBF` -->
-<!-- which may differ slightly from the one available on CRAN -->
-<!-- (until the CRAN version is updated appropriately).  -->
-
-<!-- The package in this repository can be installed from within `R` by using the following code (assuming the [devtools](https://cran.r-project.org/package=devtools)) package is available: -->
-<!-- ```R -->
-<!-- devtools::install_github("msalibian/RBF") -->
-<!-- ``` -->
-<!-- and charged in the `R` session by the following command -->
-<!-- ```{r library} -->
-<!-- library("RBF") -->
-<!-- ``` -->
-<!-- Now that the `R` package is downloaded, we can now start to see how this procedure works. -->
-
-## Boston example
-
-Consider the well-known `Boston` house price data of Harrinson and Rubinfeld (1978). This dataset was used as an example to model an additive model by Härdle et a. (2004). The data are available in the `MASS` package. It contains $n=506$ observations and 14 variables measured on the census districts of the Boston metropolitan area. Following the analysis in Härdle et a. (2004) we use the following 10 explanatory variables:
-
-- `crim`: per capita crime rate by town ($X_1$),
-- `indus`: proportion of non-retail business acres per town ($X_2$),
-- `nox`: nitric oxides concentration (parts per 10 million) ($X_3$),
-- `rm`: average number of rooms per dwelling ($X_4$),
-- `age`: proportion of owner-occupied units built prior to 1940 ($X_5$),
[…]
