### 2021-03-20 #2992 @msalibian
@mikldk We've rolled back the repository to its original state when we finished the vignette. Only the source code for the vignette (and figures which are not generated by the R Markdown file) are in the `vignettes` folder, and we removed the `doc` folder from the repository. The package passes all the usual CRAN checks, and builds (including the vignette) without any issues.

Since the HTML of the vignette and the corresponding .R file are only created (and stored in the `inst/doc` folder) when one builds the source tar ball of the package to upload to CRAN, and thus are not be immediately available to the referees from the GitHub repository, we originally decided to leave the `doc` folder and its contents (as created by `devtools::build_vignettes()`) in the repository to save the referees' some time. We then probably misread @mmrabe comment as suggesting @alemermartinez to move those files to the `vignettes` folder instead. We were aware that the `doc` folder will not be included in the package, and that the `vignettes` folder should only include the source for the vignette--we just thought we would save the referees some time.

Please let us know if you (or the referees) would like us to make the HTML version of the vignette available directly to you or them, and how. We can either temporarily restore the `doc` folder (from `devtools::build_vignettes()`) in the repository, as we had it before, or attach the HTML file to a comment here, or some other way.
