### 2021-03-22 #2992 @mikldk
@msalibian You can also try the `[pkgdown](https://pkgdown.r-lib.org/)` package which generates a webpage with the man pages and vignettes. It can be published via GitHub Pages (see the `README`).
