### 2021-03-22 #2992 @msalibian
@mikldk Done. Please find it at [https://msalibian.github.io/RBF/articles/Examples.html](https://msalibian.github.io/RBF/articles/Examples.html). Note that at this point we are only planning to use that site to help with the revision of the paper.
