### 2021-03-22 3ead0c4
change figures folder for pkgdown to work
