### 2021-03-24 #2992 @mikldk
@mcavs, @mmrabe:  Can you confirm that you have finished the review and recommend that this paper is now published?

@alemermartinez :

* In general be sure that all versions are correct and that the repo is in the state that should be published.
* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
