### 2021-03-24 #2992 @mmrabe
Yes, all issues have been resolved. I can now recommend this paper for publication in JOSS.
