### 2021-03-24 1d43d70
added missing DOIs
