+++ refs.bib
@@ -27,6 +27,7 @@
   edition = {2nd},
   publisher = {Chapman \& Hall},
   address = {London},
+  doi = {10.1201/9781351072304}
 }
 
 
@@ -57,7 +58,8 @@
   author={Maronna, Ricardo A. and Martin, R. Douglas and Yohai, Victor J. and Salibian-Barrera, Matias},
   year={2018},
   edition = {2nd},
-  publisher={John Wiley \& Sons}
+  publisher={John Wiley \& Sons},
+  doi = {10.1002/9781119214656}
 }
 
 @Manual{gam,
