+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: RBF
 Type: Package
 Title: Robust Backfitting 
 Version: 2.1.0
-Date: 2021-03-04
+Date: 2021-03-24
 Authors@R: c(person("Matias", "Salibian-Barrera", role = c("aut", "cre"), email = "matias@stat.ubc.ca"),
              person("Alejandra", "Martinez", role=c("aut"), email="ale_m_martinez@hotmail.com")
             )
