### 2021-03-25 1ec64fc
typos and minor edits
