+++ paper.md
@@ -349,7 +349,7 @@ and produced estimates very similar to the classical ones applied to the \lq\lq 
 ![Plots of estimated curves and partial residuals. 
 The solid blue lines indicate the robust fit computed 
 on the whole data set, while the classical estimators 
-computed on the \lq\lq clean\rq\rq
+computed on the \lq\lq clean\rq\rq\
 data are shown with dashed magenta lines. 
 Larger red circles indicate potential outliers. 
 \label{fig:ozoneout}](Figure-ozone-out-cla-rob.png){ width=85% }
