### 2021-03-25 #2992 @alemermartinez
Thank you very much, @mikldk<https://github.com/mikldk> @mmrabe<https://github.com/mmrabe> @mcavs<https://github.com/mcavs>

Kind regards,

Alejandra

-------- Mensaje original --------
De: Matias Salibian Barrera ***@***.***>
Fecha: 24/3/21 13:01 (GMT-03:00)
A: openjournals/joss-reviews ***@***.***>
CC: Alejandra Martinez ***@***.***>, Mention ***@***.***>
Asunto: Re: [openjournals/joss-reviews] [REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2992)

@mikldk<https://github.com/mikldk> @mmrabe<https://github.com/mmrabe> @mcavs<https://github.com/mcavs> Thank you all for your work and constructive revision process!
