### 2021-03-25 #2992 @msalibian
@mikldk The release is v2.1.0 and the DOI is [10.5281/zenodo.4637759](https://doi.org/10.5281/zenodo.4637759).
