### 2021-03-29 #2992 @mikldk
@msalibian @alemermartinez Please make the title at Zenodo the same as the paper's ("RBF: An R package to compute a robust backfitting estimator for additive models").
