### 2021-04-06 #2992 @mikldk
@danielskatz Thanks; it was a mistake; I wasn't entirely ready. Sorry.

@alemermartinez @msalibian:

1. Maybe add citation to RBF in "Statement of Need" (similar to that in the Summary; Boente et al.(2017)?)
2. Background: "Note that if [...]" for all g_j ?
3. Background: Distribution (variance? constant?) of error term (at least required for it being a standard linear regression model)?
4. Background: min formula: sigma not defined; you mention it as "residual scale", but not clear how/what it should be for the optimisation?
5. Illustration: `degree = 1 and type='Tukey'`; add spaces around second `=`. Similarly later `type='Huber'`.
