### 2021-04-06 3c5b7b9
comment on empirical version of the population optimization problem
