+++ paper.md
@@ -6,7 +6,7 @@ tags:
 - Backfitting
 - Additive models
 - Robustness
-date: "22 February 2021"
+date: "6 April 2021"
 output:
   pdf_document: default
   word_document: default
@@ -35,19 +35,17 @@ many explanatory variables. Additive models provide an alternative that is
 more flexible than linear models, not affected by the curse of
 dimensionality, and also allow the exploration of individual covariate
 effects. Standard algorithms to fit these models can be highly susceptible to
-the presence of a few atypical or outlying observations in the data.
-The ``RBF`` [@RBF] package for R implements the robust estimator for additive
+the presence of a few atypical or outlying observations in the data. The
+``RBF`` [@RBF] package for R implements the robust estimator for additive
 models of @BoenteMartinezSalibian2017, which can resist the damaging effect
 of outliers in the training set.
 
 # Statement of Need
 
-The purpose of ``RBF`` is to provide a user-friendly implementation of the 
-robust kernel-based estimation procedure for additive models 
-proposed in @BoenteMartinezSalibian2017, which are 
-resistant to
-the presence of potentially atypical or outlying observations
-in the training set. 
+The purpose of ``RBF`` is to provide a user-friendly implementation of the
+robust kernel-based estimation procedure for additive models proposed in
+@BoenteMartinezSalibian2017, which is resistant to the presence of
+potentially atypical or outlying observations in the training set.
 
 # Implementation Goals
 
@@ -84,21 +82,19 @@ variables increases, neighbourhoods rapidly become sparse, and much fewer
 training observations are available to estimate the regression function at
 any one point.
 
-If $Y$ denotes the response variable, and $\textbf{X} = (X_1, \ldots, X_d)^\top$ 
-a vector of explanatory variables, then an additive regression model
-postulates that
+If $Y$ denotes the response variable, and $\textbf{X} = (X_1, \ldots,
+X_d)^\top$ a vector of explanatory variables, then an additive regression
+model postulates that
 \begin{equation} \label{eq:model} 
 Y \ = \ \mu + \sum_{j=1}^d g_j(X_j) \, + \, \sigma \, 
 \epsilon \, ,
 \end{equation}
-where the error $\epsilon$ is independent of $\textbf{X}$ and its distribution is centered at
-zero, $\sigma > 0$ is an unknown scale parameter, 
-the location parameter $\mu \in \mathbb{R}$, and $g_j \, : \, \mathbb{R}
-\to \mathbb{R}$ are smooth functions. Note that if 
-for all $1 \le j \le d$ we have 
-$g_j( X_j ) = \beta_j \,
-X_j$ for some $\beta_j \in \mathbb{R}$, then \autoref{eq:model} reduces to a
-standard linear regression model.
+where the error $\epsilon$ is independent of $\textbf{X}$ and its
+distribution is centered at zero, $\sigma > 0$ is an unknown scale parameter,
+the location parameter $\mu \in \mathbb{R}$, and $g_j \, : \, \mathbb{R} \to
+\mathbb{R}$ are smooth functions. Note that if for all $1 \le j \le d$ we
+have $g_j( X_j ) = \beta_j \, X_j$ for some $\beta_j \in \mathbb{R}$, then
+\autoref{eq:model} reduces to a standard linear regression model.
 <!-- The functions $g_j$ can be interpreted as the marginal effect
 of the $j$-th covariate on the expected value of the response when all
 other explanatory variables remain fixed.-->
@@ -110,8 +106,8 @@ components $g_j$. It is based on the following observation: under
 \sum_{\ell \ne j} g_\ell(X_\ell) | X_j = x ]$. Thus, each $g_j$ is iteratively
 computed by smoothing the partial residuals as functions of $X_j$.
 
[…]
