### 2021-04-06 4a99d0b
updated date and preliminary hat(sigma) description
