### 2021-04-06 #2992 @msalibian
@mikldk Thank you for catching these omissions. We've fixed them. For comments (3) and (4), we now explicitly include the residual scale parameter in our model in equation (1), and we also mention that \sigma is estimated using a preliminary estimator (the MAD of the residuals from a local median fit). The distribution family of the errors in equation (1) may remain unspecified, but we mention that it is assumed to be symmetric around zero.
