### 2021-04-07 #2992 @danielskatz
I've found a few small things that shoud be changed, as indicated in https://github.com/msalibian/RBF/pull/6
