### 2021-04-07 #2992 @alemermartinez
Yes, he is. Thank you, Daniel, for this new read of the text.

Kind regards,

Alejandra

-------- Mensaje original --------
De: "Daniel S. Katz" ***@***.***>
Fecha: 7/4/21 17:57 (GMT-03:00)
A: openjournals/joss-reviews ***@***.***>
CC: Alejandra Martinez ***@***.***>, Mention ***@***.***>
Asunto: Re: [openjournals/joss-reviews] [REVIEW]: RBF: An R package to compute a robust backfitting estimator for additive models (#2992)

Are you speaking for @alemermartinez<https://github.com/alemermartinez> as well?
